@@ -1,5 +1,5 @@
 module trpc.group/trpc/trpc-protocol/pb/go/trpc/v2
 
-go 1.16
+go 1.23.8
 
 require google.golang.org/protobuf v1.36.6
