@@ -23,7 +23,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.36.6
 // 	protoc        v5.27.1
 // source: trpc/api/annotations.proto
 
@@ -34,6 +34,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	descriptorpb "google.golang.org/protobuf/types/descriptorpb"
 	reflect "reflect"
+	unsafe "unsafe"
 )
 
 const (
@@ -76,34 +77,14 @@ var (
 
 var File_trpc_api_annotations_proto protoreflect.FileDescriptor
 
-var file_trpc_api_annotations_proto_rawDesc = []byte{
-	0x0a, 0x1a, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x61, 0x6e, 0x6e, 0x6f, 0x74,
-	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x08, 0x74, 0x72,
-	0x70, 0x63, 0x2e, 0x61, 0x70, 0x69, 0x1a, 0x13, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x61, 0x70, 0x69,
-	0x2f, 0x68, 0x74, 0x74, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x18, 0x74, 0x72, 0x70,
-	0x63, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x6f, 0x70, 0x65, 0x6e, 0x61, 0x70, 0x69, 0x76, 0x32, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x20, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x64, 0x65, 0x73, 0x63, 0x72, 0x69, 0x70, 0x74, 0x6f,
-	0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x3a, 0x48, 0x0a, 0x04, 0x68, 0x74, 0x74, 0x70, 0x12,
-	0x1e, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2e, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x18,
-	0x99, 0x88, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x12, 0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x61,
-	0x70, 0x69, 0x2e, 0x48, 0x74, 0x74, 0x70, 0x52, 0x75, 0x6c, 0x65, 0x52, 0x04, 0x68, 0x74, 0x74,
-	0x70, 0x3a, 0x5e, 0x0a, 0x10, 0x6f, 0x70, 0x65, 0x6e, 0x61, 0x70, 0x69, 0x76, 0x32, 0x5f, 0x73,
-	0x63, 0x68, 0x65, 0x6d, 0x61, 0x12, 0x1f, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x4f,
-	0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x18, 0x99, 0x88, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10,
-	0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x63, 0x68, 0x65, 0x6d, 0x61,
-	0x52, 0x0f, 0x6f, 0x70, 0x65, 0x6e, 0x61, 0x70, 0x69, 0x76, 0x32, 0x53, 0x63, 0x68, 0x65, 0x6d,
-	0x61, 0x42, 0x4d, 0x0a, 0x1d, 0x63, 0x6f, 0x6d, 0x2e, 0x74, 0x65, 0x6e, 0x63, 0x65, 0x6e, 0x74,
-	0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x65,
-	0x78, 0x74, 0x5a, 0x2c, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x67, 0x72, 0x6f, 0x75, 0x70, 0x2f, 0x74,
-	0x72, 0x70, 0x63, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f,
-	0x6c, 0x2f, 0x70, 0x62, 0x2f, 0x67, 0x6f, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x61, 0x70, 0x69,
-	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_trpc_api_annotations_proto_rawDesc = "" +
+	"\n" +
+	"\x1atrpc/api/annotations.proto\x12\btrpc.api\x1a\x13trpc/api/http.proto\x1a\x18trpc/api/openapiv2.proto\x1a google/protobuf/descriptor.proto:H\n" +
+	"\x04http\x12\x1e.google.protobuf.MethodOptions\x18\x99\x88\x03 \x01(\v2\x12.trpc.api.HttpRuleR\x04http:^\n" +
+	"\x10openapiv2_schema\x12\x1f.google.protobuf.MessageOptions\x18\x99\x88\x03 \x01(\v2\x10.trpc.api.SchemaR\x0fopenapiv2SchemaBM\n" +
+	"\x1dcom.tencent.trpc.protobuf.extZ,trpc.group/trpc/trpc-protocol/pb/go/trpc/apib\x06proto3"
 
-var file_trpc_api_annotations_proto_goTypes = []interface{}{
+var file_trpc_api_annotations_proto_goTypes = []any{
 	(*descriptorpb.MethodOptions)(nil),  // 0: google.protobuf.MethodOptions
 	(*descriptorpb.MessageOptions)(nil), // 1: google.protobuf.MessageOptions
 	(*HttpRule)(nil),                    // 2: trpc.api.HttpRule
@@ -132,7 +113,7 @@ func file_trpc_api_annotations_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_trpc_api_annotations_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_trpc_api_annotations_proto_rawDesc), len(file_trpc_api_annotations_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   0,
 			NumExtensions: 2,
@@ -143,7 +124,6 @@ func file_trpc_api_annotations_proto_init() {
 		ExtensionInfos:    file_trpc_api_annotations_proto_extTypes,
 	}.Build()
 	File_trpc_api_annotations_proto = out.File
-	file_trpc_api_annotations_proto_rawDesc = nil
 	file_trpc_api_annotations_proto_goTypes = nil
 	file_trpc_api_annotations_proto_depIdxs = nil
 }
