@@ -6,7 +6,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.36.6
 // 	protoc        v5.27.1
 // source: trpc/api/openapiv2.proto
 
@@ -17,6 +17,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -28,20 +29,17 @@ const (
 
 // message option schema
 type Schema struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	JsonSchema    *JSONSchema            `protobuf:"bytes,1,opt,name=json_schema,json=jsonSchema,proto3" json:"json_schema,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	JsonSchema *JSONSchema `protobuf:"bytes,1,opt,name=json_schema,json=jsonSchema,proto3" json:"json_schema,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Schema) Reset() {
 	*x = Schema{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_api_openapiv2_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_api_openapiv2_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Schema) String() string {
@@ -52,7 +50,7 @@ func (*Schema) ProtoMessage() {}
 
 func (x *Schema) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_api_openapiv2_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -75,22 +73,20 @@ func (x *Schema) GetJsonSchema() *JSONSchema {
 }
 
 // `JSONSchema` represents properties from JSON SCHEMA taken,and as used, in
-//  the OpenAPI v2 spec.
+//
+//	the OpenAPI v2 spec.
 type JSONSchema struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Required      []string               `protobuf:"bytes,1,rep,name=required,proto3" json:"required,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Required []string `protobuf:"bytes,1,rep,name=required,proto3" json:"required,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *JSONSchema) Reset() {
 	*x = JSONSchema{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_api_openapiv2_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_api_openapiv2_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *JSONSchema) String() string {
@@ -101,7 +97,7 @@ func (*JSONSchema) ProtoMessage() {}
 
 func (x *JSONSchema) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_api_openapiv2_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -125,38 +121,31 @@ func (x *JSONSchema) GetRequired() []string {
 
 var File_trpc_api_openapiv2_proto protoreflect.FileDescriptor
 
-var file_trpc_api_openapiv2_proto_rawDesc = []byte{
-	0x0a, 0x18, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x6f, 0x70, 0x65, 0x6e, 0x61,
-	0x70, 0x69, 0x76, 0x32, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x08, 0x74, 0x72, 0x70, 0x63,
-	0x2e, 0x61, 0x70, 0x69, 0x22, 0x3f, 0x0a, 0x06, 0x53, 0x63, 0x68, 0x65, 0x6d, 0x61, 0x12, 0x35,
-	0x0a, 0x0b, 0x6a, 0x73, 0x6f, 0x6e, 0x5f, 0x73, 0x63, 0x68, 0x65, 0x6d, 0x61, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x14, 0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x4a,
-	0x53, 0x4f, 0x4e, 0x53, 0x63, 0x68, 0x65, 0x6d, 0x61, 0x52, 0x0a, 0x6a, 0x73, 0x6f, 0x6e, 0x53,
-	0x63, 0x68, 0x65, 0x6d, 0x61, 0x22, 0x28, 0x0a, 0x0a, 0x4a, 0x53, 0x4f, 0x4e, 0x53, 0x63, 0x68,
-	0x65, 0x6d, 0x61, 0x12, 0x1a, 0x0a, 0x08, 0x72, 0x65, 0x71, 0x75, 0x69, 0x72, 0x65, 0x64, 0x18,
-	0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x08, 0x72, 0x65, 0x71, 0x75, 0x69, 0x72, 0x65, 0x64, 0x42,
-	0x4d, 0x0a, 0x1d, 0x63, 0x6f, 0x6d, 0x2e, 0x74, 0x65, 0x6e, 0x63, 0x65, 0x6e, 0x74, 0x2e, 0x74,
-	0x72, 0x70, 0x63, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x65, 0x78, 0x74,
-	0x5a, 0x2c, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x67, 0x72, 0x6f, 0x75, 0x70, 0x2f, 0x74, 0x72, 0x70,
-	0x63, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x2f,
-	0x70, 0x62, 0x2f, 0x67, 0x6f, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x61, 0x70, 0x69, 0x62, 0x06,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_trpc_api_openapiv2_proto_rawDesc = "" +
+	"\n" +
+	"\x18trpc/api/openapiv2.proto\x12\btrpc.api\"?\n" +
+	"\x06Schema\x125\n" +
+	"\vjson_schema\x18\x01 \x01(\v2\x14.trpc.api.JSONSchemaR\n" +
+	"jsonSchema\"(\n" +
+	"\n" +
+	"JSONSchema\x12\x1a\n" +
+	"\brequired\x18\x01 \x03(\tR\brequiredBM\n" +
+	"\x1dcom.tencent.trpc.protobuf.extZ,trpc.group/trpc/trpc-protocol/pb/go/trpc/apib\x06proto3"
 
 var (
 	file_trpc_api_openapiv2_proto_rawDescOnce sync.Once
-	file_trpc_api_openapiv2_proto_rawDescData = file_trpc_api_openapiv2_proto_rawDesc
+	file_trpc_api_openapiv2_proto_rawDescData []byte
 )
 
 func file_trpc_api_openapiv2_proto_rawDescGZIP() []byte {
 	file_trpc_api_openapiv2_proto_rawDescOnce.Do(func() {
-		file_trpc_api_openapiv2_proto_rawDescData = protoimpl.X.CompressGZIP(file_trpc_api_openapiv2_proto_rawDescData)
+		file_trpc_api_openapiv2_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_trpc_api_openapiv2_proto_rawDesc), len(file_trpc_api_openapiv2_proto_rawDesc)))
 	})
 	return file_trpc_api_openapiv2_proto_rawDescData
 }
 
 var file_trpc_api_openapiv2_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_trpc_api_openapiv2_proto_goTypes = []interface{}{
+var file_trpc_api_openapiv2_proto_goTypes = []any{
 	(*Schema)(nil),     // 0: trpc.api.Schema
 	(*JSONSchema)(nil), // 1: trpc.api.JSONSchema
 }
@@ -174,37 +163,11 @@ func file_trpc_api_openapiv2_proto_init() {
 	if File_trpc_api_openapiv2_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_trpc_api_openapiv2_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Schema); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_api_openapiv2_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*JSONSchema); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_trpc_api_openapiv2_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_trpc_api_openapiv2_proto_rawDesc), len(file_trpc_api_openapiv2_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -215,7 +178,6 @@ func file_trpc_api_openapiv2_proto_init() {
 		MessageInfos:      file_trpc_api_openapiv2_proto_msgTypes,
 	}.Build()
 	File_trpc_api_openapiv2_proto = out.File
-	file_trpc_api_openapiv2_proto_rawDesc = nil
 	file_trpc_api_openapiv2_proto_goTypes = nil
 	file_trpc_api_openapiv2_proto_depIdxs = nil
 }
