@@ -6,7 +6,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.36.6
 // 	protoc        v5.27.1
 // source: trpc/trpc.proto
 
@@ -17,6 +17,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -796,10 +797,7 @@ func (TrpcRetCode) EnumDescriptor() ([]byte, []int) {
 
 // The message definition of stream `INIT` frame
 type TrpcStreamInitMeta struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// request meta information
 	RequestMeta *TrpcStreamInitRequestMeta `protobuf:"bytes,1,opt,name=request_meta,json=requestMeta,proto3" json:"request_meta,omitempty"`
 	// response meta information
@@ -814,15 +812,15 @@ type TrpcStreamInitMeta struct {
 	// eg: gzip/snappy/..., not used by default
 	// The specific value corresponds to `TrpcCompressType`
 	ContentEncoding uint32 `protobuf:"varint,5,opt,name=content_encoding,json=contentEncoding,proto3" json:"content_encoding,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *TrpcStreamInitMeta) Reset() {
 	*x = TrpcStreamInitMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_trpc_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_trpc_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TrpcStreamInitMeta) String() string {
@@ -833,7 +831,7 @@ func (*TrpcStreamInitMeta) ProtoMessage() {}
 
 func (x *TrpcStreamInitMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_trpc_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -885,10 +883,7 @@ func (x *TrpcStreamInitMeta) GetContentEncoding() uint32 {
 
 // The request meta information definition of stream `INIT` frame
 type TrpcStreamInitRequestMeta struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Caller name
 	// The specification format: trpc.application_name.server_name.proto_service_name, 4 segments
 	Caller []byte `protobuf:"bytes,1,opt,name=caller,proto3" json:"caller,omitempty"`
@@ -905,20 +900,20 @@ type TrpcStreamInitRequestMeta struct {
 	// The information key-value pair transparently transmitted by the framework
 	// Currently divided into two parts:
 	// 1 part is the information to be transparently transmitted by the framework layer,
-	// and the name of the key must be started with `trpc-``
+	// and the name of the key must be started with `trpc-“
 	// 2 part is the information to be transparently transmitted by the business layer,
-	// and the business can set it by itself, it is recommended to start with `app-``, not `trpc-`
+	// and the business can set it by itself, it is recommended to start with `app-“, not `trpc-`
 	// Note: The key-value pair in trans_info will be transparently transmitted through the whole link, please use it carefully for business.
-	TransInfo map[string][]byte `protobuf:"bytes,5,rep,name=trans_info,json=transInfo,proto3" json:"trans_info,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	TransInfo     map[string][]byte `protobuf:"bytes,5,rep,name=trans_info,json=transInfo,proto3" json:"trans_info,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TrpcStreamInitRequestMeta) Reset() {
 	*x = TrpcStreamInitRequestMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_trpc_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_trpc_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TrpcStreamInitRequestMeta) String() string {
@@ -929,7 +924,7 @@ func (*TrpcStreamInitRequestMeta) ProtoMessage() {}
 
 func (x *TrpcStreamInitRequestMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_trpc_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -981,24 +976,21 @@ func (x *TrpcStreamInitRequestMeta) GetTransInfo() map[string][]byte {
 
 // The response meta information definition of stream `INIT` frame
 type TrpcStreamInitResponseMeta struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Error code
 	// The specific value corresponds to `TrpcRetCode`
 	Ret int32 `protobuf:"varint,1,opt,name=ret,proto3" json:"ret,omitempty"`
 	// The result information when the call fails
-	ErrorMsg []byte `protobuf:"bytes,2,opt,name=error_msg,json=errorMsg,proto3" json:"error_msg,omitempty"`
+	ErrorMsg      []byte `protobuf:"bytes,2,opt,name=error_msg,json=errorMsg,proto3" json:"error_msg,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TrpcStreamInitResponseMeta) Reset() {
 	*x = TrpcStreamInitResponseMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_trpc_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_trpc_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TrpcStreamInitResponseMeta) String() string {
@@ -1009,7 +1001,7 @@ func (*TrpcStreamInitResponseMeta) ProtoMessage() {}
 
 func (x *TrpcStreamInitResponseMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_trpc_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1040,21 +1032,18 @@ func (x *TrpcStreamInitResponseMeta) GetErrorMsg() []byte {
 
 // The meta information definition of stream `FEEDBACK` frame
 type TrpcStreamFeedBackMeta struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// increased window size
 	WindowSizeIncrement uint32 `protobuf:"varint,1,opt,name=window_size_increment,json=windowSizeIncrement,proto3" json:"window_size_increment,omitempty"`
+	unknownFields       protoimpl.UnknownFields
+	sizeCache           protoimpl.SizeCache
 }
 
 func (x *TrpcStreamFeedBackMeta) Reset() {
 	*x = TrpcStreamFeedBackMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_trpc_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_trpc_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TrpcStreamFeedBackMeta) String() string {
@@ -1065,7 +1054,7 @@ func (*TrpcStreamFeedBackMeta) ProtoMessage() {}
 
 func (x *TrpcStreamFeedBackMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_trpc_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1089,10 +1078,7 @@ func (x *TrpcStreamFeedBackMeta) GetWindowSizeIncrement() uint32 {
 
 // The meta information definition of trpc stream protocol for closing stream
 type TrpcStreamCloseMeta struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// The type of stream closure, close one end, or close all
 	CloseType int32 `protobuf:"varint,1,opt,name=close_type,json=closeType,proto3" json:"close_type,omitempty"`
 	// Error code
@@ -1107,23 +1093,23 @@ type TrpcStreamCloseMeta struct {
 	// The information key-value pair transparently transmitted by the framework
 	// Currently divided into two parts:
 	// 1 part is the information to be transparently transmitted by the framework layer,
-	// and the name of the key must be started with `trpc-``
+	// and the name of the key must be started with `trpc-“
 	// 2 part is the information to be transparently transmitted by the business layer,
-	// and the business can set it by itself, it is recommended to start with `app-``, not `trpc-`
+	// and the business can set it by itself, it is recommended to start with `app-“, not `trpc-`
 	// Note: The key-value pair in trans_info will be transparently transmitted through the whole link, please use it carefully for business.
-	TransInfo map[string][]byte `protobuf:"bytes,5,rep,name=trans_info,json=transInfo,proto3" json:"trans_info,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	TransInfo map[string][]byte `protobuf:"bytes,5,rep,name=trans_info,json=transInfo,proto3" json:"trans_info,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	// The error code of the interface
 	// 0 means success, other means failure
-	FuncRet int32 `protobuf:"varint,6,opt,name=func_ret,json=funcRet,proto3" json:"func_ret,omitempty"`
+	FuncRet       int32 `protobuf:"varint,6,opt,name=func_ret,json=funcRet,proto3" json:"func_ret,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TrpcStreamCloseMeta) Reset() {
 	*x = TrpcStreamCloseMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_trpc_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_trpc_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TrpcStreamCloseMeta) String() string {
@@ -1134,7 +1120,7 @@ func (*TrpcStreamCloseMeta) ProtoMessage() {}
 
 func (x *TrpcStreamCloseMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_trpc_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1193,10 +1179,7 @@ func (x *TrpcStreamCloseMeta) GetFuncRet() int32 {
 
 // The request header for unary
 type RequestProtocol struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// The version of protocol
 	// The specific value corresponds to `TrpcProtoVersion`
 	Version uint32 `protobuf:"varint,1,opt,name=version,proto3" json:"version,omitempty"`
@@ -1224,11 +1207,11 @@ type RequestProtocol struct {
 	// The information key-value pair transparently transmitted by the framework
 	// Currently divided into two parts:
 	// 1 part is the information to be transparently transmitted by the framework layer,
-	// and the name of the key must be started with `trpc-``
+	// and the name of the key must be started with `trpc-“
 	// 2 part is the information to be transparently transmitted by the business layer,
-	// and the business can set it by itself, it is recommended to start with `app-``, not `trpc-`
+	// and the business can set it by itself, it is recommended to start with `app-“, not `trpc-`
 	// Note: The key-value pair in trans_info will be transparently transmitted through the whole link, please use it carefully for business.
-	TransInfo map[string][]byte `protobuf:"bytes,9,rep,name=trans_info,json=transInfo,proto3" json:"trans_info,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	TransInfo map[string][]byte `protobuf:"bytes,9,rep,name=trans_info,json=transInfo,proto3" json:"trans_info,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	// The serialization type of the request data
 	// eg: proto/json/.., default proto
 	// The specific value corresponds to `TrpcContentEncodeType`
@@ -1239,15 +1222,15 @@ type RequestProtocol struct {
 	ContentEncoding uint32 `protobuf:"varint,11,opt,name=content_encoding,json=contentEncoding,proto3" json:"content_encoding,omitempty"`
 	// The size of attachment data
 	AttachmentSize uint32 `protobuf:"varint,12,opt,name=attachment_size,json=attachmentSize,proto3" json:"attachment_size,omitempty"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *RequestProtocol) Reset() {
 	*x = RequestProtocol{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_trpc_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_trpc_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RequestProtocol) String() string {
@@ -1258,7 +1241,7 @@ func (*RequestProtocol) ProtoMessage() {}
 
 func (x *RequestProtocol) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_trpc_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1359,10 +1342,7 @@ func (x *RequestProtocol) GetAttachmentSize() uint32 {
 
 // The response header for unary
 type ResponseProtocol struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// The version of protocol
 	// The specific value corresponds to `TrpcProtoVersion`
 	Version uint32 `protobuf:"varint,1,opt,name=version,proto3" json:"version,omitempty"`
@@ -1387,10 +1367,10 @@ type ResponseProtocol struct {
 	// The information key-value pair transparently transmitted by the framework
 	// Currently divided into two parts:
 	// 1 part is the information to be transparently transmitted by the framework layer,
-	// and the name of the key must be started with `trpc-``
+	// and the name of the key must be started with `trpc-“
 	// 2 part is the information to be transparently transmitted by the business layer,
-	// and the business can set it by itself, it is recommended to start with `app-``, not `trpc-`
-	TransInfo map[string][]byte `protobuf:"bytes,8,rep,name=trans_info,json=transInfo,proto3" json:"trans_info,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	// and the business can set it by itself, it is recommended to start with `app-“, not `trpc-`
+	TransInfo map[string][]byte `protobuf:"bytes,8,rep,name=trans_info,json=transInfo,proto3" json:"trans_info,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	// The serialization type of the request data
 	// eg: proto/json/.., default proto
 	// The specific value corresponds to `TrpcContentEncodeType`
@@ -1401,15 +1381,15 @@ type ResponseProtocol struct {
 	ContentEncoding uint32 `protobuf:"varint,10,opt,name=content_encoding,json=contentEncoding,proto3" json:"content_encoding,omitempty"`
 	// The size of attachment data
 	AttachmentSize uint32 `protobuf:"varint,12,opt,name=attachment_size,json=attachmentSize,proto3" json:"attachment_size,omitempty"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *ResponseProtocol) Reset() {
 	*x = ResponseProtocol{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_trpc_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_trpc_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ResponseProtocol) String() string {
@@ -1420,7 +1400,7 @@ func (*ResponseProtocol) ProtoMessage() {}
 
 func (x *ResponseProtocol) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_trpc_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1514,338 +1494,196 @@ func (x *ResponseProtocol) GetAttachmentSize() uint32 {
 
 var File_trpc_trpc_proto protoreflect.FileDescriptor
 
-var file_trpc_trpc_proto_rawDesc = []byte{
-	0x0a, 0x0f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x12, 0x04, 0x74, 0x72, 0x70, 0x63, 0x22, 0x97, 0x02, 0x0a, 0x12, 0x54, 0x72, 0x70, 0x63,
-	0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49, 0x6e, 0x69, 0x74, 0x4d, 0x65, 0x74, 0x61, 0x12, 0x42,
-	0x0a, 0x0c, 0x72, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x5f, 0x6d, 0x65, 0x74, 0x61, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1f, 0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x54, 0x72, 0x70, 0x63,
-	0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49, 0x6e, 0x69, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x4d, 0x65, 0x74, 0x61, 0x52, 0x0b, 0x72, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x4d, 0x65,
-	0x74, 0x61, 0x12, 0x45, 0x0a, 0x0d, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x5f, 0x6d,
-	0x65, 0x74, 0x61, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x20, 0x2e, 0x74, 0x72, 0x70, 0x63,
-	0x2e, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49, 0x6e, 0x69, 0x74, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x4d, 0x65, 0x74, 0x61, 0x52, 0x0c, 0x72, 0x65, 0x73,
-	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x4d, 0x65, 0x74, 0x61, 0x12, 0x28, 0x0a, 0x10, 0x69, 0x6e, 0x69,
-	0x74, 0x5f, 0x77, 0x69, 0x6e, 0x64, 0x6f, 0x77, 0x5f, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x0d, 0x52, 0x0e, 0x69, 0x6e, 0x69, 0x74, 0x57, 0x69, 0x6e, 0x64, 0x6f, 0x77, 0x53,
-	0x69, 0x7a, 0x65, 0x12, 0x21, 0x0a, 0x0c, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x5f, 0x74,
-	0x79, 0x70, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x63, 0x6f, 0x6e, 0x74, 0x65,
-	0x6e, 0x74, 0x54, 0x79, 0x70, 0x65, 0x12, 0x29, 0x0a, 0x10, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e,
-	0x74, 0x5f, 0x65, 0x6e, 0x63, 0x6f, 0x64, 0x69, 0x6e, 0x67, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0d,
-	0x52, 0x0f, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x45, 0x6e, 0x63, 0x6f, 0x64, 0x69, 0x6e,
-	0x67, 0x22, 0x8f, 0x02, 0x0a, 0x19, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
-	0x49, 0x6e, 0x69, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x4d, 0x65, 0x74, 0x61, 0x12,
-	0x16, 0x0a, 0x06, 0x63, 0x61, 0x6c, 0x6c, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52,
-	0x06, 0x63, 0x61, 0x6c, 0x6c, 0x65, 0x72, 0x12, 0x16, 0x0a, 0x06, 0x63, 0x61, 0x6c, 0x6c, 0x65,
-	0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x06, 0x63, 0x61, 0x6c, 0x6c, 0x65, 0x65, 0x12,
-	0x12, 0x0a, 0x04, 0x66, 0x75, 0x6e, 0x63, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x66,
-	0x75, 0x6e, 0x63, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x5f, 0x74,
-	0x79, 0x70, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x4d, 0x0a, 0x0a, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x5f,
-	0x69, 0x6e, 0x66, 0x6f, 0x18, 0x05, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x2e, 0x2e, 0x74, 0x72, 0x70,
-	0x63, 0x2e, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49, 0x6e, 0x69, 0x74,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x4d, 0x65, 0x74, 0x61, 0x2e, 0x54, 0x72, 0x61, 0x6e,
-	0x73, 0x49, 0x6e, 0x66, 0x6f, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x09, 0x74, 0x72, 0x61, 0x6e,
-	0x73, 0x49, 0x6e, 0x66, 0x6f, 0x1a, 0x3c, 0x0a, 0x0e, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x49, 0x6e,
-	0x66, 0x6f, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x14, 0x0a, 0x05, 0x76, 0x61, 0x6c,
-	0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a,
-	0x02, 0x38, 0x01, 0x22, 0x4b, 0x0a, 0x1a, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65, 0x61,
-	0x6d, 0x49, 0x6e, 0x69, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x4d, 0x65, 0x74,
-	0x61, 0x12, 0x10, 0x0a, 0x03, 0x72, 0x65, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x03,
-	0x72, 0x65, 0x74, 0x12, 0x1b, 0x0a, 0x09, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x5f, 0x6d, 0x73, 0x67,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x08, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x4d, 0x73, 0x67,
-	0x22, 0x4c, 0x0a, 0x16, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x65,
-	0x65, 0x64, 0x42, 0x61, 0x63, 0x6b, 0x4d, 0x65, 0x74, 0x61, 0x12, 0x32, 0x0a, 0x15, 0x77, 0x69,
-	0x6e, 0x64, 0x6f, 0x77, 0x5f, 0x73, 0x69, 0x7a, 0x65, 0x5f, 0x69, 0x6e, 0x63, 0x72, 0x65, 0x6d,
-	0x65, 0x6e, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x13, 0x77, 0x69, 0x6e, 0x64, 0x6f,
-	0x77, 0x53, 0x69, 0x7a, 0x65, 0x49, 0x6e, 0x63, 0x72, 0x65, 0x6d, 0x65, 0x6e, 0x74, 0x22, 0x9d,
-	0x02, 0x0a, 0x13, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x43, 0x6c, 0x6f,
-	0x73, 0x65, 0x4d, 0x65, 0x74, 0x61, 0x12, 0x1d, 0x0a, 0x0a, 0x63, 0x6c, 0x6f, 0x73, 0x65, 0x5f,
-	0x74, 0x79, 0x70, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x09, 0x63, 0x6c, 0x6f, 0x73,
-	0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x72, 0x65, 0x74, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x05, 0x52, 0x03, 0x72, 0x65, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x6d, 0x73, 0x67, 0x18, 0x03,
-	0x20, 0x01, 0x28, 0x0c, 0x52, 0x03, 0x6d, 0x73, 0x67, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x65, 0x73,
-	0x73, 0x61, 0x67, 0x65, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0d, 0x52,
-	0x0b, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x47, 0x0a, 0x0a,
-	0x74, 0x72, 0x61, 0x6e, 0x73, 0x5f, 0x69, 0x6e, 0x66, 0x6f, 0x18, 0x05, 0x20, 0x03, 0x28, 0x0b,
-	0x32, 0x28, 0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65,
-	0x61, 0x6d, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x4d, 0x65, 0x74, 0x61, 0x2e, 0x54, 0x72, 0x61, 0x6e,
-	0x73, 0x49, 0x6e, 0x66, 0x6f, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x09, 0x74, 0x72, 0x61, 0x6e,
-	0x73, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x19, 0x0a, 0x08, 0x66, 0x75, 0x6e, 0x63, 0x5f, 0x72, 0x65,
-	0x74, 0x18, 0x06, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x66, 0x75, 0x6e, 0x63, 0x52, 0x65, 0x74,
-	0x1a, 0x3c, 0x0a, 0x0e, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x49, 0x6e, 0x66, 0x6f, 0x45, 0x6e, 0x74,
-	0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x03, 0x6b, 0x65, 0x79, 0x12, 0x14, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x0c, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02, 0x38, 0x01, 0x22, 0xe2,
-	0x03, 0x0a, 0x0f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x63,
-	0x6f, 0x6c, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x0d, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x1b, 0x0a, 0x09,
-	0x63, 0x61, 0x6c, 0x6c, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52,
-	0x08, 0x63, 0x61, 0x6c, 0x6c, 0x54, 0x79, 0x70, 0x65, 0x12, 0x1d, 0x0a, 0x0a, 0x72, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x5f, 0x69, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x09, 0x72,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x49, 0x64, 0x12, 0x18, 0x0a, 0x07, 0x74, 0x69, 0x6d, 0x65,
-	0x6f, 0x75, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x07, 0x74, 0x69, 0x6d, 0x65, 0x6f,
-	0x75, 0x74, 0x12, 0x16, 0x0a, 0x06, 0x63, 0x61, 0x6c, 0x6c, 0x65, 0x72, 0x18, 0x05, 0x20, 0x01,
-	0x28, 0x0c, 0x52, 0x06, 0x63, 0x61, 0x6c, 0x6c, 0x65, 0x72, 0x12, 0x16, 0x0a, 0x06, 0x63, 0x61,
-	0x6c, 0x6c, 0x65, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x06, 0x63, 0x61, 0x6c, 0x6c,
-	0x65, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x66, 0x75, 0x6e, 0x63, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0c,
-	0x52, 0x04, 0x66, 0x75, 0x6e, 0x63, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x6d, 0x65,
-	0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x43, 0x0a, 0x0a, 0x74, 0x72, 0x61,
-	0x6e, 0x73, 0x5f, 0x69, 0x6e, 0x66, 0x6f, 0x18, 0x09, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x24, 0x2e,
-	0x74, 0x72, 0x70, 0x63, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x74,
-	0x6f, 0x63, 0x6f, 0x6c, 0x2e, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x49, 0x6e, 0x66, 0x6f, 0x45, 0x6e,
-	0x74, 0x72, 0x79, 0x52, 0x09, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x21,
-	0x0a, 0x0c, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x18, 0x0a,
-	0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x54, 0x79, 0x70,
-	0x65, 0x12, 0x29, 0x0a, 0x10, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x5f, 0x65, 0x6e, 0x63,
-	0x6f, 0x64, 0x69, 0x6e, 0x67, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0f, 0x63, 0x6f, 0x6e,
-	0x74, 0x65, 0x6e, 0x74, 0x45, 0x6e, 0x63, 0x6f, 0x64, 0x69, 0x6e, 0x67, 0x12, 0x27, 0x0a, 0x0f,
-	0x61, 0x74, 0x74, 0x61, 0x63, 0x68, 0x6d, 0x65, 0x6e, 0x74, 0x5f, 0x73, 0x69, 0x7a, 0x65, 0x18,
-	0x0c, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0e, 0x61, 0x74, 0x74, 0x61, 0x63, 0x68, 0x6d, 0x65, 0x6e,
-	0x74, 0x53, 0x69, 0x7a, 0x65, 0x1a, 0x3c, 0x0a, 0x0e, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x49, 0x6e,
-	0x66, 0x6f, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x14, 0x0a, 0x05, 0x76, 0x61, 0x6c,
-	0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a,
-	0x02, 0x38, 0x01, 0x22, 0xd0, 0x03, 0x0a, 0x10, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
-	0x50, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73,
-	0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69,
-	0x6f, 0x6e, 0x12, 0x1b, 0x0a, 0x09, 0x63, 0x61, 0x6c, 0x6c, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x08, 0x63, 0x61, 0x6c, 0x6c, 0x54, 0x79, 0x70, 0x65, 0x12,
-	0x1d, 0x0a, 0x0a, 0x72, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x5f, 0x69, 0x64, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x0d, 0x52, 0x09, 0x72, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x49, 0x64, 0x12, 0x10,
-	0x0a, 0x03, 0x72, 0x65, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x05, 0x52, 0x03, 0x72, 0x65, 0x74,
-	0x12, 0x19, 0x0a, 0x08, 0x66, 0x75, 0x6e, 0x63, 0x5f, 0x72, 0x65, 0x74, 0x18, 0x05, 0x20, 0x01,
-	0x28, 0x05, 0x52, 0x07, 0x66, 0x75, 0x6e, 0x63, 0x52, 0x65, 0x74, 0x12, 0x1b, 0x0a, 0x09, 0x65,
-	0x72, 0x72, 0x6f, 0x72, 0x5f, 0x6d, 0x73, 0x67, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x08,
-	0x65, 0x72, 0x72, 0x6f, 0x72, 0x4d, 0x73, 0x67, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x65, 0x73, 0x73,
-	0x61, 0x67, 0x65, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b,
-	0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x44, 0x0a, 0x0a, 0x74,
-	0x72, 0x61, 0x6e, 0x73, 0x5f, 0x69, 0x6e, 0x66, 0x6f, 0x18, 0x08, 0x20, 0x03, 0x28, 0x0b, 0x32,
-	0x25, 0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x50,
-	0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x2e, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x49, 0x6e, 0x66,
-	0x6f, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x09, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x49, 0x6e, 0x66,
-	0x6f, 0x12, 0x21, 0x0a, 0x0c, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x5f, 0x74, 0x79, 0x70,
-	0x65, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74,
-	0x54, 0x79, 0x70, 0x65, 0x12, 0x29, 0x0a, 0x10, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x5f,
-	0x65, 0x6e, 0x63, 0x6f, 0x64, 0x69, 0x6e, 0x67, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0f,
-	0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x45, 0x6e, 0x63, 0x6f, 0x64, 0x69, 0x6e, 0x67, 0x12,
-	0x27, 0x0a, 0x0f, 0x61, 0x74, 0x74, 0x61, 0x63, 0x68, 0x6d, 0x65, 0x6e, 0x74, 0x5f, 0x73, 0x69,
-	0x7a, 0x65, 0x18, 0x0c, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0e, 0x61, 0x74, 0x74, 0x61, 0x63, 0x68,
-	0x6d, 0x65, 0x6e, 0x74, 0x53, 0x69, 0x7a, 0x65, 0x1a, 0x3c, 0x0a, 0x0e, 0x54, 0x72, 0x61, 0x6e,
-	0x73, 0x49, 0x6e, 0x66, 0x6f, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65,
-	0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x14, 0x0a, 0x05,
-	0x76, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x05, 0x76, 0x61, 0x6c,
-	0x75, 0x65, 0x3a, 0x02, 0x38, 0x01, 0x2a, 0x39, 0x0a, 0x09, 0x54, 0x72, 0x70, 0x63, 0x4d, 0x61,
-	0x67, 0x69, 0x63, 0x12, 0x15, 0x0a, 0x11, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x44, 0x45, 0x46, 0x41,
-	0x55, 0x4c, 0x54, 0x5f, 0x4e, 0x4f, 0x4e, 0x45, 0x10, 0x00, 0x12, 0x15, 0x0a, 0x10, 0x54, 0x52,
-	0x50, 0x43, 0x5f, 0x4d, 0x41, 0x47, 0x49, 0x43, 0x5f, 0x56, 0x41, 0x4c, 0x55, 0x45, 0x10, 0xb0,
-	0x12, 0x2a, 0x40, 0x0a, 0x11, 0x54, 0x72, 0x70, 0x63, 0x44, 0x61, 0x74, 0x61, 0x46, 0x72, 0x61,
-	0x6d, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x14, 0x0a, 0x10, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x55,
-	0x4e, 0x41, 0x52, 0x59, 0x5f, 0x46, 0x52, 0x41, 0x4d, 0x45, 0x10, 0x00, 0x12, 0x15, 0x0a, 0x11,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x46, 0x52, 0x41, 0x4d,
-	0x45, 0x10, 0x01, 0x2a, 0x9a, 0x01, 0x0a, 0x13, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65,
-	0x61, 0x6d, 0x46, 0x72, 0x61, 0x6d, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x0e, 0x0a, 0x0a, 0x54,
-	0x52, 0x50, 0x43, 0x5f, 0x55, 0x4e, 0x41, 0x52, 0x59, 0x10, 0x00, 0x12, 0x1a, 0x0a, 0x16, 0x54,
-	0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x46, 0x52, 0x41, 0x4d, 0x45,
-	0x5f, 0x49, 0x4e, 0x49, 0x54, 0x10, 0x01, 0x12, 0x1a, 0x0a, 0x16, 0x54, 0x52, 0x50, 0x43, 0x5f,
-	0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x46, 0x52, 0x41, 0x4d, 0x45, 0x5f, 0x44, 0x41, 0x54,
-	0x41, 0x10, 0x02, 0x12, 0x1e, 0x0a, 0x1a, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45,
-	0x41, 0x4d, 0x5f, 0x46, 0x52, 0x41, 0x4d, 0x45, 0x5f, 0x46, 0x45, 0x45, 0x44, 0x42, 0x41, 0x43,
-	0x4b, 0x10, 0x03, 0x12, 0x1b, 0x0a, 0x17, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45,
-	0x41, 0x4d, 0x5f, 0x46, 0x52, 0x41, 0x4d, 0x45, 0x5f, 0x43, 0x4c, 0x4f, 0x53, 0x45, 0x10, 0x04,
-	0x2a, 0x43, 0x0a, 0x13, 0x54, 0x72, 0x70, 0x63, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x43, 0x6c,
-	0x6f, 0x73, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x15, 0x0a, 0x11, 0x54, 0x52, 0x50, 0x43, 0x5f,
-	0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x4f, 0x53, 0x45, 0x10, 0x00, 0x12, 0x15,
-	0x0a, 0x11, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x52, 0x45,
-	0x53, 0x45, 0x54, 0x10, 0x01, 0x2a, 0x25, 0x0a, 0x10, 0x54, 0x72, 0x70, 0x63, 0x50, 0x72, 0x6f,
-	0x74, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x11, 0x0a, 0x0d, 0x54, 0x52, 0x50,
-	0x43, 0x5f, 0x50, 0x52, 0x4f, 0x54, 0x4f, 0x5f, 0x56, 0x31, 0x10, 0x00, 0x2a, 0x39, 0x0a, 0x0c,
-	0x54, 0x72, 0x70, 0x63, 0x43, 0x61, 0x6c, 0x6c, 0x54, 0x79, 0x70, 0x65, 0x12, 0x13, 0x0a, 0x0f,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x55, 0x4e, 0x41, 0x52, 0x59, 0x5f, 0x43, 0x41, 0x4c, 0x4c, 0x10,
-	0x00, 0x12, 0x14, 0x0a, 0x10, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x4f, 0x4e, 0x45, 0x57, 0x41, 0x59,
-	0x5f, 0x43, 0x41, 0x4c, 0x4c, 0x10, 0x01, 0x2a, 0xa1, 0x01, 0x0a, 0x0f, 0x54, 0x72, 0x70, 0x63,
-	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x10, 0x0a, 0x0c, 0x54,
-	0x52, 0x50, 0x43, 0x5f, 0x44, 0x45, 0x46, 0x41, 0x55, 0x4c, 0x54, 0x10, 0x00, 0x12, 0x17, 0x0a,
-	0x13, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x44, 0x59, 0x45, 0x49, 0x4e, 0x47, 0x5f, 0x4d, 0x45, 0x53,
-	0x53, 0x41, 0x47, 0x45, 0x10, 0x01, 0x12, 0x16, 0x0a, 0x12, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x54,
-	0x52, 0x41, 0x43, 0x45, 0x5f, 0x4d, 0x45, 0x53, 0x53, 0x41, 0x47, 0x45, 0x10, 0x02, 0x12, 0x1a,
-	0x0a, 0x16, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x4d, 0x55, 0x4c, 0x54, 0x49, 0x5f, 0x45, 0x4e, 0x56,
-	0x5f, 0x4d, 0x45, 0x53, 0x53, 0x41, 0x47, 0x45, 0x10, 0x04, 0x12, 0x15, 0x0a, 0x11, 0x54, 0x52,
-	0x50, 0x43, 0x5f, 0x47, 0x52, 0x49, 0x44, 0x5f, 0x4d, 0x45, 0x53, 0x53, 0x41, 0x47, 0x45, 0x10,
-	0x08, 0x12, 0x18, 0x0a, 0x14, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x45, 0x54, 0x4e, 0x41, 0x4d,
-	0x45, 0x5f, 0x4d, 0x45, 0x53, 0x53, 0x41, 0x47, 0x45, 0x10, 0x10, 0x2a, 0xf2, 0x01, 0x0a, 0x15,
-	0x54, 0x72, 0x70, 0x63, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x45, 0x6e, 0x63, 0x6f, 0x64,
-	0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x15, 0x0a, 0x11, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x50, 0x52,
-	0x4f, 0x54, 0x4f, 0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44, 0x45, 0x10, 0x00, 0x12, 0x13, 0x0a, 0x0f,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x4a, 0x43, 0x45, 0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44, 0x45, 0x10,
-	0x01, 0x12, 0x14, 0x0a, 0x10, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x4a, 0x53, 0x4f, 0x4e, 0x5f, 0x45,
-	0x4e, 0x43, 0x4f, 0x44, 0x45, 0x10, 0x02, 0x12, 0x1a, 0x0a, 0x16, 0x54, 0x52, 0x50, 0x43, 0x5f,
-	0x46, 0x4c, 0x41, 0x54, 0x42, 0x55, 0x46, 0x46, 0x45, 0x52, 0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44,
-	0x45, 0x10, 0x03, 0x12, 0x14, 0x0a, 0x10, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x4e, 0x4f, 0x4f, 0x50,
-	0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44, 0x45, 0x10, 0x04, 0x12, 0x13, 0x0a, 0x0f, 0x54, 0x52, 0x50,
-	0x43, 0x5f, 0x58, 0x4d, 0x4c, 0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44, 0x45, 0x10, 0x05, 0x12, 0x16,
-	0x0a, 0x12, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x54, 0x48, 0x52, 0x49, 0x46, 0x54, 0x5f, 0x45, 0x4e,
-	0x43, 0x4f, 0x44, 0x45, 0x10, 0x06, 0x12, 0x1e, 0x0a, 0x1a, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x54,
-	0x48, 0x52, 0x49, 0x46, 0x54, 0x5f, 0x43, 0x4f, 0x4d, 0x50, 0x41, 0x43, 0x54, 0x5f, 0x45, 0x4e,
-	0x43, 0x4f, 0x44, 0x45, 0x10, 0x07, 0x12, 0x18, 0x0a, 0x14, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x54,
-	0x45, 0x58, 0x54, 0x5f, 0x58, 0x4d, 0x4c, 0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44, 0x45, 0x10, 0x08,
-	0x2a, 0xf2, 0x01, 0x0a, 0x10, 0x54, 0x72, 0x70, 0x63, 0x43, 0x6f, 0x6d, 0x70, 0x72, 0x65, 0x73,
-	0x73, 0x54, 0x79, 0x70, 0x65, 0x12, 0x19, 0x0a, 0x15, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x44, 0x45,
-	0x46, 0x41, 0x55, 0x4c, 0x54, 0x5f, 0x43, 0x4f, 0x4d, 0x50, 0x52, 0x45, 0x53, 0x53, 0x10, 0x00,
-	0x12, 0x16, 0x0a, 0x12, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x47, 0x5a, 0x49, 0x50, 0x5f, 0x43, 0x4f,
-	0x4d, 0x50, 0x52, 0x45, 0x53, 0x53, 0x10, 0x01, 0x12, 0x18, 0x0a, 0x14, 0x54, 0x52, 0x50, 0x43,
-	0x5f, 0x53, 0x4e, 0x41, 0x50, 0x50, 0x59, 0x5f, 0x43, 0x4f, 0x4d, 0x50, 0x52, 0x45, 0x53, 0x53,
-	0x10, 0x02, 0x12, 0x16, 0x0a, 0x12, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x5a, 0x4c, 0x49, 0x42, 0x5f,
-	0x43, 0x4f, 0x4d, 0x50, 0x52, 0x45, 0x53, 0x53, 0x10, 0x03, 0x12, 0x1f, 0x0a, 0x1b, 0x54, 0x52,
-	0x50, 0x43, 0x5f, 0x53, 0x4e, 0x41, 0x50, 0x50, 0x59, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d,
-	0x5f, 0x43, 0x4f, 0x4d, 0x50, 0x52, 0x45, 0x53, 0x53, 0x10, 0x04, 0x12, 0x1e, 0x0a, 0x1a, 0x54,
-	0x52, 0x50, 0x43, 0x5f, 0x53, 0x4e, 0x41, 0x50, 0x50, 0x59, 0x5f, 0x42, 0x4c, 0x4f, 0x43, 0x4b,
-	0x5f, 0x43, 0x4f, 0x4d, 0x50, 0x52, 0x45, 0x53, 0x53, 0x10, 0x05, 0x12, 0x1b, 0x0a, 0x17, 0x54,
-	0x52, 0x50, 0x43, 0x5f, 0x4c, 0x5a, 0x34, 0x5f, 0x46, 0x52, 0x41, 0x4d, 0x45, 0x5f, 0x43, 0x4f,
-	0x4d, 0x50, 0x52, 0x45, 0x53, 0x53, 0x10, 0x06, 0x12, 0x1b, 0x0a, 0x17, 0x54, 0x52, 0x50, 0x43,
-	0x5f, 0x4c, 0x5a, 0x34, 0x5f, 0x42, 0x4c, 0x4f, 0x43, 0x4b, 0x5f, 0x43, 0x4f, 0x4d, 0x50, 0x52,
-	0x45, 0x53, 0x53, 0x10, 0x07, 0x2a, 0xa5, 0x0e, 0x0a, 0x0b, 0x54, 0x72, 0x70, 0x63, 0x52, 0x65,
-	0x74, 0x43, 0x6f, 0x64, 0x65, 0x12, 0x17, 0x0a, 0x13, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x49, 0x4e,
-	0x56, 0x4f, 0x4b, 0x45, 0x5f, 0x53, 0x55, 0x43, 0x43, 0x45, 0x53, 0x53, 0x10, 0x00, 0x12, 0x1a,
-	0x0a, 0x16, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x44, 0x45,
-	0x43, 0x4f, 0x44, 0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x01, 0x12, 0x1a, 0x0a, 0x16, 0x54, 0x52,
-	0x50, 0x43, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44, 0x45,
-	0x5f, 0x45, 0x52, 0x52, 0x10, 0x02, 0x12, 0x1d, 0x0a, 0x19, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53,
-	0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x4e, 0x4f, 0x53, 0x45, 0x52, 0x56, 0x49, 0x43, 0x45, 0x5f,
-	0x45, 0x52, 0x52, 0x10, 0x0b, 0x12, 0x1a, 0x0a, 0x16, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x45,
-	0x52, 0x56, 0x45, 0x52, 0x5f, 0x4e, 0x4f, 0x46, 0x55, 0x4e, 0x43, 0x5f, 0x45, 0x52, 0x52, 0x10,
-	0x0c, 0x12, 0x1b, 0x0a, 0x17, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52,
-	0x5f, 0x54, 0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x15, 0x12, 0x1c,
-	0x0a, 0x18, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x4f, 0x56,
-	0x45, 0x52, 0x4c, 0x4f, 0x41, 0x44, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x16, 0x12, 0x1b, 0x0a, 0x17,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x4c, 0x49, 0x4d, 0x49,
-	0x54, 0x45, 0x44, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x17, 0x12, 0x25, 0x0a, 0x21, 0x54, 0x52, 0x50,
-	0x43, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x46, 0x55, 0x4c, 0x4c, 0x5f, 0x4c, 0x49,
-	0x4e, 0x4b, 0x5f, 0x54, 0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x18,
-	0x12, 0x1a, 0x0a, 0x16, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f,
-	0x53, 0x59, 0x53, 0x54, 0x45, 0x4d, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x1f, 0x12, 0x18, 0x0a, 0x14,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x41, 0x55, 0x54, 0x48,
-	0x5f, 0x45, 0x52, 0x52, 0x10, 0x29, 0x12, 0x1c, 0x0a, 0x18, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53,
-	0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x56, 0x41, 0x4c, 0x49, 0x44, 0x41, 0x54, 0x45, 0x5f, 0x45,
-	0x52, 0x52, 0x10, 0x33, 0x12, 0x22, 0x0a, 0x1e, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x43, 0x4c, 0x49,
-	0x45, 0x4e, 0x54, 0x5f, 0x49, 0x4e, 0x56, 0x4f, 0x4b, 0x45, 0x5f, 0x54, 0x49, 0x4d, 0x45, 0x4f,
-	0x55, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x65, 0x12, 0x25, 0x0a, 0x21, 0x54, 0x52, 0x50, 0x43,
-	0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x46, 0x55, 0x4c, 0x4c, 0x5f, 0x4c, 0x49, 0x4e,
-	0x4b, 0x5f, 0x54, 0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x66, 0x12,
-	0x1b, 0x0a, 0x17, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x43,
-	0x4f, 0x4e, 0x4e, 0x45, 0x43, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x6f, 0x12, 0x1a, 0x0a, 0x16,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x45, 0x4e, 0x43, 0x4f,
-	0x44, 0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x79, 0x12, 0x1a, 0x0a, 0x16, 0x54, 0x52, 0x50, 0x43,
-	0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x44, 0x45, 0x43, 0x4f, 0x44, 0x45, 0x5f, 0x45,
-	0x52, 0x52, 0x10, 0x7a, 0x12, 0x1b, 0x0a, 0x17, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x43, 0x4c, 0x49,
-	0x45, 0x4e, 0x54, 0x5f, 0x4c, 0x49, 0x4d, 0x49, 0x54, 0x45, 0x44, 0x5f, 0x45, 0x52, 0x52, 0x10,
-	0x7b, 0x12, 0x1c, 0x0a, 0x18, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54,
-	0x5f, 0x4f, 0x56, 0x45, 0x52, 0x4c, 0x4f, 0x41, 0x44, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x7c, 0x12,
-	0x1b, 0x0a, 0x16, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x52,
-	0x4f, 0x55, 0x54, 0x45, 0x52, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x83, 0x01, 0x12, 0x1c, 0x0a, 0x17,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x4e, 0x45, 0x54, 0x57,
-	0x4f, 0x52, 0x4b, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x8d, 0x01, 0x12, 0x1d, 0x0a, 0x18, 0x54, 0x52,
-	0x50, 0x43, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x56, 0x41, 0x4c, 0x49, 0x44, 0x41,
-	0x54, 0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0x97, 0x01, 0x12, 0x1d, 0x0a, 0x18, 0x54, 0x52, 0x50,
-	0x43, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x43, 0x41, 0x4e, 0x43, 0x45, 0x4c, 0x45,
-	0x44, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xa1, 0x01, 0x12, 0x1f, 0x0a, 0x1a, 0x54, 0x52, 0x50, 0x43,
-	0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x52, 0x45, 0x41, 0x44, 0x5f, 0x46, 0x52, 0x41,
-	0x4d, 0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xab, 0x01, 0x12, 0x23, 0x0a, 0x1e, 0x54, 0x52, 0x50,
-	0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f,
-	0x4e, 0x45, 0x54, 0x57, 0x4f, 0x52, 0x4b, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xc9, 0x01, 0x12, 0x2c,
-	0x0a, 0x27, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45,
-	0x52, 0x56, 0x45, 0x52, 0x5f, 0x4d, 0x53, 0x47, 0x5f, 0x45, 0x58, 0x43, 0x45, 0x45, 0x44, 0x5f,
-	0x4c, 0x49, 0x4d, 0x49, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xd3, 0x01, 0x12, 0x22, 0x0a, 0x1d,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56,
-	0x45, 0x52, 0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44, 0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xdd, 0x01,
-	0x12, 0x22, 0x0a, 0x1d, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f,
-	0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x44, 0x45, 0x43, 0x4f, 0x44, 0x45, 0x5f, 0x45, 0x52,
-	0x52, 0x10, 0xde, 0x01, 0x12, 0x21, 0x0a, 0x1c, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52,
-	0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x57, 0x52, 0x49, 0x54, 0x45,
-	0x5f, 0x45, 0x4e, 0x44, 0x10, 0xe7, 0x01, 0x12, 0x2a, 0x0a, 0x25, 0x54, 0x52, 0x50, 0x43, 0x5f,
-	0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x57, 0x52,
-	0x49, 0x54, 0x45, 0x5f, 0x4f, 0x56, 0x45, 0x52, 0x46, 0x4c, 0x4f, 0x57, 0x5f, 0x45, 0x52, 0x52,
-	0x10, 0xe8, 0x01, 0x12, 0x27, 0x0a, 0x22, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45,
-	0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x57, 0x52, 0x49, 0x54, 0x45, 0x5f,
-	0x43, 0x4c, 0x4f, 0x53, 0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xe9, 0x01, 0x12, 0x29, 0x0a, 0x24,
-	0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56,
-	0x45, 0x52, 0x5f, 0x57, 0x52, 0x49, 0x54, 0x45, 0x5f, 0x54, 0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54,
-	0x5f, 0x45, 0x52, 0x52, 0x10, 0xea, 0x01, 0x12, 0x20, 0x0a, 0x1b, 0x54, 0x52, 0x50, 0x43, 0x5f,
-	0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x52, 0x45,
-	0x41, 0x44, 0x5f, 0x45, 0x4e, 0x44, 0x10, 0xfb, 0x01, 0x12, 0x26, 0x0a, 0x21, 0x54, 0x52, 0x50,
-	0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f,
-	0x52, 0x45, 0x41, 0x44, 0x5f, 0x43, 0x4c, 0x4f, 0x53, 0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xfc,
-	0x01, 0x12, 0x26, 0x0a, 0x21, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d,
-	0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x52, 0x45, 0x41, 0x44, 0x5f, 0x45, 0x4d, 0x50,
-	0x54, 0x59, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xfd, 0x01, 0x12, 0x28, 0x0a, 0x23, 0x54, 0x52, 0x50,
-	0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f,
-	0x52, 0x45, 0x41, 0x44, 0x5f, 0x54, 0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54, 0x5f, 0x45, 0x52, 0x52,
-	0x10, 0xfe, 0x01, 0x12, 0x28, 0x0a, 0x23, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45,
-	0x41, 0x4d, 0x5f, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x5f, 0x49, 0x44, 0x4c, 0x45, 0x5f, 0x54,
-	0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xff, 0x01, 0x12, 0x23, 0x0a,
-	0x1e, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49,
-	0x45, 0x4e, 0x54, 0x5f, 0x4e, 0x45, 0x54, 0x57, 0x4f, 0x52, 0x4b, 0x5f, 0x45, 0x52, 0x52, 0x10,
-	0xad, 0x02, 0x12, 0x2c, 0x0a, 0x27, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41,
-	0x4d, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x4d, 0x53, 0x47, 0x5f, 0x45, 0x58, 0x43,
-	0x45, 0x45, 0x44, 0x5f, 0x4c, 0x49, 0x4d, 0x49, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xb7, 0x02,
-	0x12, 0x22, 0x0a, 0x1d, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f,
-	0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x45, 0x4e, 0x43, 0x4f, 0x44, 0x45, 0x5f, 0x45, 0x52,
-	0x52, 0x10, 0xc1, 0x02, 0x12, 0x22, 0x0a, 0x1d, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52,
-	0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x44, 0x45, 0x43, 0x4f, 0x44,
-	0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xc2, 0x02, 0x12, 0x21, 0x0a, 0x1c, 0x54, 0x52, 0x50, 0x43,
-	0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x57,
-	0x52, 0x49, 0x54, 0x45, 0x5f, 0x45, 0x4e, 0x44, 0x10, 0xcb, 0x02, 0x12, 0x2a, 0x0a, 0x25, 0x54,
-	0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e,
-	0x54, 0x5f, 0x57, 0x52, 0x49, 0x54, 0x45, 0x5f, 0x4f, 0x56, 0x45, 0x52, 0x46, 0x4c, 0x4f, 0x57,
-	0x5f, 0x45, 0x52, 0x52, 0x10, 0xcc, 0x02, 0x12, 0x27, 0x0a, 0x22, 0x54, 0x52, 0x50, 0x43, 0x5f,
-	0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x57, 0x52,
-	0x49, 0x54, 0x45, 0x5f, 0x43, 0x4c, 0x4f, 0x53, 0x45, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xcd, 0x02,
-	0x12, 0x29, 0x0a, 0x24, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f,
-	0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x57, 0x52, 0x49, 0x54, 0x45, 0x5f, 0x54, 0x49, 0x4d,
-	0x45, 0x4f, 0x55, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xce, 0x02, 0x12, 0x20, 0x0a, 0x1b, 0x54,
-	0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e,
-	0x54, 0x5f, 0x52, 0x45, 0x41, 0x44, 0x5f, 0x45, 0x4e, 0x44, 0x10, 0xdf, 0x02, 0x12, 0x26, 0x0a,
-	0x21, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49,
-	0x45, 0x4e, 0x54, 0x5f, 0x52, 0x45, 0x41, 0x44, 0x5f, 0x43, 0x4c, 0x4f, 0x53, 0x45, 0x5f, 0x45,
-	0x52, 0x52, 0x10, 0xe0, 0x02, 0x12, 0x26, 0x0a, 0x21, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54,
-	0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x52, 0x45, 0x41, 0x44,
-	0x5f, 0x45, 0x4d, 0x50, 0x54, 0x59, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xe1, 0x02, 0x12, 0x28, 0x0a,
-	0x23, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49,
-	0x45, 0x4e, 0x54, 0x5f, 0x52, 0x45, 0x41, 0x44, 0x5f, 0x54, 0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54,
-	0x5f, 0x45, 0x52, 0x52, 0x10, 0xe2, 0x02, 0x12, 0x28, 0x0a, 0x23, 0x54, 0x52, 0x50, 0x43, 0x5f,
-	0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x43, 0x4c, 0x49, 0x45, 0x4e, 0x54, 0x5f, 0x49, 0x44,
-	0x4c, 0x45, 0x5f, 0x54, 0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xe3,
-	0x02, 0x12, 0x1c, 0x0a, 0x17, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x49, 0x4e, 0x56, 0x4f, 0x4b, 0x45,
-	0x5f, 0x55, 0x4e, 0x4b, 0x4e, 0x4f, 0x57, 0x4e, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xe7, 0x07, 0x12,
-	0x1c, 0x0a, 0x17, 0x54, 0x52, 0x50, 0x43, 0x5f, 0x53, 0x54, 0x52, 0x45, 0x41, 0x4d, 0x5f, 0x55,
-	0x4e, 0x4b, 0x4e, 0x4f, 0x57, 0x4e, 0x5f, 0x45, 0x52, 0x52, 0x10, 0xe8, 0x07, 0x42, 0x60, 0x0a,
-	0x26, 0x63, 0x6f, 0x6d, 0x2e, 0x74, 0x65, 0x6e, 0x63, 0x65, 0x6e, 0x74, 0x2e, 0x74, 0x72, 0x70,
-	0x63, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x73, 0x74, 0x61, 0x6e, 0x64, 0x61, 0x72, 0x64,
-	0x2e, 0x63, 0x6f, 0x6d, 0x6d, 0x6f, 0x6e, 0x42, 0x0c, 0x54, 0x52, 0x50, 0x43, 0x50, 0x72, 0x6f,
-	0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x5a, 0x28, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x67, 0x72, 0x6f, 0x75,
-	0x70, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2d, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x63, 0x6f, 0x6c, 0x2f, 0x70, 0x62, 0x2f, 0x67, 0x6f, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x62,
-	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_trpc_trpc_proto_rawDesc = "" +
+	"\n" +
+	"\x0ftrpc/trpc.proto\x12\x04trpc\"\x97\x02\n" +
+	"\x12TrpcStreamInitMeta\x12B\n" +
+	"\frequest_meta\x18\x01 \x01(\v2\x1f.trpc.TrpcStreamInitRequestMetaR\vrequestMeta\x12E\n" +
+	"\rresponse_meta\x18\x02 \x01(\v2 .trpc.TrpcStreamInitResponseMetaR\fresponseMeta\x12(\n" +
+	"\x10init_window_size\x18\x03 \x01(\rR\x0einitWindowSize\x12!\n" +
+	"\fcontent_type\x18\x04 \x01(\rR\vcontentType\x12)\n" +
+	"\x10content_encoding\x18\x05 \x01(\rR\x0fcontentEncoding\"\x8f\x02\n" +
+	"\x19TrpcStreamInitRequestMeta\x12\x16\n" +
+	"\x06caller\x18\x01 \x01(\fR\x06caller\x12\x16\n" +
+	"\x06callee\x18\x02 \x01(\fR\x06callee\x12\x12\n" +
+	"\x04func\x18\x03 \x01(\fR\x04func\x12!\n" +
+	"\fmessage_type\x18\x04 \x01(\rR\vmessageType\x12M\n" +
+	"\n" +
+	"trans_info\x18\x05 \x03(\v2..trpc.TrpcStreamInitRequestMeta.TransInfoEntryR\ttransInfo\x1a<\n" +
+	"\x0eTransInfoEntry\x12\x10\n" +
+	"\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n" +
+	"\x05value\x18\x02 \x01(\fR\x05value:\x028\x01\"K\n" +
+	"\x1aTrpcStreamInitResponseMeta\x12\x10\n" +
+	"\x03ret\x18\x01 \x01(\x05R\x03ret\x12\x1b\n" +
+	"\terror_msg\x18\x02 \x01(\fR\berrorMsg\"L\n" +
+	"\x16TrpcStreamFeedBackMeta\x122\n" +
+	"\x15window_size_increment\x18\x01 \x01(\rR\x13windowSizeIncrement\"\x9d\x02\n" +
+	"\x13TrpcStreamCloseMeta\x12\x1d\n" +
+	"\n" +
+	"close_type\x18\x01 \x01(\x05R\tcloseType\x12\x10\n" +
+	"\x03ret\x18\x02 \x01(\x05R\x03ret\x12\x10\n" +
+	"\x03msg\x18\x03 \x01(\fR\x03msg\x12!\n" +
+	"\fmessage_type\x18\x04 \x01(\rR\vmessageType\x12G\n" +
+	"\n" +
+	"trans_info\x18\x05 \x03(\v2(.trpc.TrpcStreamCloseMeta.TransInfoEntryR\ttransInfo\x12\x19\n" +
+	"\bfunc_ret\x18\x06 \x01(\x05R\afuncRet\x1a<\n" +
+	"\x0eTransInfoEntry\x12\x10\n" +
+	"\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n" +
+	"\x05value\x18\x02 \x01(\fR\x05value:\x028\x01\"\xe2\x03\n" +
+	"\x0fRequestProtocol\x12\x18\n" +
+	"\aversion\x18\x01 \x01(\rR\aversion\x12\x1b\n" +
+	"\tcall_type\x18\x02 \x01(\rR\bcallType\x12\x1d\n" +
+	"\n" +
+	"request_id\x18\x03 \x01(\rR\trequestId\x12\x18\n" +
+	"\atimeout\x18\x04 \x01(\rR\atimeout\x12\x16\n" +
+	"\x06caller\x18\x05 \x01(\fR\x06caller\x12\x16\n" +
+	"\x06callee\x18\x06 \x01(\fR\x06callee\x12\x12\n" +
+	"\x04func\x18\a \x01(\fR\x04func\x12!\n" +
+	"\fmessage_type\x18\b \x01(\rR\vmessageType\x12C\n" +
+	"\n" +
+	"trans_info\x18\t \x03(\v2$.trpc.RequestProtocol.TransInfoEntryR\ttransInfo\x12!\n" +
+	"\fcontent_type\x18\n" +
+	" \x01(\rR\vcontentType\x12)\n" +
+	"\x10content_encoding\x18\v \x01(\rR\x0fcontentEncoding\x12'\n" +
+	"\x0fattachment_size\x18\f \x01(\rR\x0eattachmentSize\x1a<\n" +
+	"\x0eTransInfoEntry\x12\x10\n" +
+	"\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n" +
+	"\x05value\x18\x02 \x01(\fR\x05value:\x028\x01\"\xd0\x03\n" +
+	"\x10ResponseProtocol\x12\x18\n" +
+	"\aversion\x18\x01 \x01(\rR\aversion\x12\x1b\n" +
+	"\tcall_type\x18\x02 \x01(\rR\bcallType\x12\x1d\n" +
+	"\n" +
+	"request_id\x18\x03 \x01(\rR\trequestId\x12\x10\n" +
+	"\x03ret\x18\x04 \x01(\x05R\x03ret\x12\x19\n" +
+	"\bfunc_ret\x18\x05 \x01(\x05R\afuncRet\x12\x1b\n" +
+	"\terror_msg\x18\x06 \x01(\fR\berrorMsg\x12!\n" +
+	"\fmessage_type\x18\a \x01(\rR\vmessageType\x12D\n" +
+	"\n" +
+	"trans_info\x18\b \x03(\v2%.trpc.ResponseProtocol.TransInfoEntryR\ttransInfo\x12!\n" +
+	"\fcontent_type\x18\t \x01(\rR\vcontentType\x12)\n" +
+	"\x10content_encoding\x18\n" +
+	" \x01(\rR\x0fcontentEncoding\x12'\n" +
+	"\x0fattachment_size\x18\f \x01(\rR\x0eattachmentSize\x1a<\n" +
+	"\x0eTransInfoEntry\x12\x10\n" +
+	"\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n" +
+	"\x05value\x18\x02 \x01(\fR\x05value:\x028\x01*9\n" +
+	"\tTrpcMagic\x12\x15\n" +
+	"\x11TRPC_DEFAULT_NONE\x10\x00\x12\x15\n" +
+	"\x10TRPC_MAGIC_VALUE\x10\xb0\x12*@\n" +
+	"\x11TrpcDataFrameType\x12\x14\n" +
+	"\x10TRPC_UNARY_FRAME\x10\x00\x12\x15\n" +
+	"\x11TRPC_STREAM_FRAME\x10\x01*\x9a\x01\n" +
+	"\x13TrpcStreamFrameType\x12\x0e\n" +
+	"\n" +
+	"TRPC_UNARY\x10\x00\x12\x1a\n" +
+	"\x16TRPC_STREAM_FRAME_INIT\x10\x01\x12\x1a\n" +
+	"\x16TRPC_STREAM_FRAME_DATA\x10\x02\x12\x1e\n" +
+	"\x1aTRPC_STREAM_FRAME_FEEDBACK\x10\x03\x12\x1b\n" +
+	"\x17TRPC_STREAM_FRAME_CLOSE\x10\x04*C\n" +
+	"\x13TrpcStreamCloseType\x12\x15\n" +
+	"\x11TRPC_STREAM_CLOSE\x10\x00\x12\x15\n" +
+	"\x11TRPC_STREAM_RESET\x10\x01*%\n" +
+	"\x10TrpcProtoVersion\x12\x11\n" +
+	"\rTRPC_PROTO_V1\x10\x00*9\n" +
+	"\fTrpcCallType\x12\x13\n" +
+	"\x0fTRPC_UNARY_CALL\x10\x00\x12\x14\n" +
+	"\x10TRPC_ONEWAY_CALL\x10\x01*\xa1\x01\n" +
+	"\x0fTrpcMessageType\x12\x10\n" +
+	"\fTRPC_DEFAULT\x10\x00\x12\x17\n" +
+	"\x13TRPC_DYEING_MESSAGE\x10\x01\x12\x16\n" +
+	"\x12TRPC_TRACE_MESSAGE\x10\x02\x12\x1a\n" +
+	"\x16TRPC_MULTI_ENV_MESSAGE\x10\x04\x12\x15\n" +
+	"\x11TRPC_GRID_MESSAGE\x10\b\x12\x18\n" +
+	"\x14TRPC_SETNAME_MESSAGE\x10\x10*\xf2\x01\n" +
+	"\x15TrpcContentEncodeType\x12\x15\n" +
+	"\x11TRPC_PROTO_ENCODE\x10\x00\x12\x13\n" +
+	"\x0fTRPC_JCE_ENCODE\x10\x01\x12\x14\n" +
+	"\x10TRPC_JSON_ENCODE\x10\x02\x12\x1a\n" +
+	"\x16TRPC_FLATBUFFER_ENCODE\x10\x03\x12\x14\n" +
+	"\x10TRPC_NOOP_ENCODE\x10\x04\x12\x13\n" +
+	"\x0fTRPC_XML_ENCODE\x10\x05\x12\x16\n" +
+	"\x12TRPC_THRIFT_ENCODE\x10\x06\x12\x1e\n" +
+	"\x1aTRPC_THRIFT_COMPACT_ENCODE\x10\a\x12\x18\n" +
+	"\x14TRPC_TEXT_XML_ENCODE\x10\b*\xf2\x01\n" +
+	"\x10TrpcCompressType\x12\x19\n" +
+	"\x15TRPC_DEFAULT_COMPRESS\x10\x00\x12\x16\n" +
+	"\x12TRPC_GZIP_COMPRESS\x10\x01\x12\x18\n" +
+	"\x14TRPC_SNAPPY_COMPRESS\x10\x02\x12\x16\n" +
+	"\x12TRPC_ZLIB_COMPRESS\x10\x03\x12\x1f\n" +
+	"\x1bTRPC_SNAPPY_STREAM_COMPRESS\x10\x04\x12\x1e\n" +
+	"\x1aTRPC_SNAPPY_BLOCK_COMPRESS\x10\x05\x12\x1b\n" +
+	"\x17TRPC_LZ4_FRAME_COMPRESS\x10\x06\x12\x1b\n" +
+	"\x17TRPC_LZ4_BLOCK_COMPRESS\x10\a*\xa5\x0e\n" +
+	"\vTrpcRetCode\x12\x17\n" +
+	"\x13TRPC_INVOKE_SUCCESS\x10\x00\x12\x1a\n" +
+	"\x16TRPC_SERVER_DECODE_ERR\x10\x01\x12\x1a\n" +
+	"\x16TRPC_SERVER_ENCODE_ERR\x10\x02\x12\x1d\n" +
+	"\x19TRPC_SERVER_NOSERVICE_ERR\x10\v\x12\x1a\n" +
+	"\x16TRPC_SERVER_NOFUNC_ERR\x10\f\x12\x1b\n" +
+	"\x17TRPC_SERVER_TIMEOUT_ERR\x10\x15\x12\x1c\n" +
+	"\x18TRPC_SERVER_OVERLOAD_ERR\x10\x16\x12\x1b\n" +
+	"\x17TRPC_SERVER_LIMITED_ERR\x10\x17\x12%\n" +
+	"!TRPC_SERVER_FULL_LINK_TIMEOUT_ERR\x10\x18\x12\x1a\n" +
+	"\x16TRPC_SERVER_SYSTEM_ERR\x10\x1f\x12\x18\n" +
+	"\x14TRPC_SERVER_AUTH_ERR\x10)\x12\x1c\n" +
+	"\x18TRPC_SERVER_VALIDATE_ERR\x103\x12\"\n" +
+	"\x1eTRPC_CLIENT_INVOKE_TIMEOUT_ERR\x10e\x12%\n" +
+	"!TRPC_CLIENT_FULL_LINK_TIMEOUT_ERR\x10f\x12\x1b\n" +
+	"\x17TRPC_CLIENT_CONNECT_ERR\x10o\x12\x1a\n" +
+	"\x16TRPC_CLIENT_ENCODE_ERR\x10y\x12\x1a\n" +
+	"\x16TRPC_CLIENT_DECODE_ERR\x10z\x12\x1b\n" +
+	"\x17TRPC_CLIENT_LIMITED_ERR\x10{\x12\x1c\n" +
+	"\x18TRPC_CLIENT_OVERLOAD_ERR\x10|\x12\x1b\n" +
+	"\x16TRPC_CLIENT_ROUTER_ERR\x10\x83\x01\x12\x1c\n" +
+	"\x17TRPC_CLIENT_NETWORK_ERR\x10\x8d\x01\x12\x1d\n" +
+	"\x18TRPC_CLIENT_VALIDATE_ERR\x10\x97\x01\x12\x1d\n" +
+	"\x18TRPC_CLIENT_CANCELED_ERR\x10\xa1\x01\x12\x1f\n" +
+	"\x1aTRPC_CLIENT_READ_FRAME_ERR\x10\xab\x01\x12#\n" +
+	"\x1eTRPC_STREAM_SERVER_NETWORK_ERR\x10\xc9\x01\x12,\n" +
+	"'TRPC_STREAM_SERVER_MSG_EXCEED_LIMIT_ERR\x10\xd3\x01\x12\"\n" +
+	"\x1dTRPC_STREAM_SERVER_ENCODE_ERR\x10\xdd\x01\x12\"\n" +
+	"\x1dTRPC_STREAM_SERVER_DECODE_ERR\x10\xde\x01\x12!\n" +
+	"\x1cTRPC_STREAM_SERVER_WRITE_END\x10\xe7\x01\x12*\n" +
+	"%TRPC_STREAM_SERVER_WRITE_OVERFLOW_ERR\x10\xe8\x01\x12'\n" +
+	"\"TRPC_STREAM_SERVER_WRITE_CLOSE_ERR\x10\xe9\x01\x12)\n" +
+	"$TRPC_STREAM_SERVER_WRITE_TIMEOUT_ERR\x10\xea\x01\x12 \n" +
+	"\x1bTRPC_STREAM_SERVER_READ_END\x10\xfb\x01\x12&\n" +
+	"!TRPC_STREAM_SERVER_READ_CLOSE_ERR\x10\xfc\x01\x12&\n" +
+	"!TRPC_STREAM_SERVER_READ_EMPTY_ERR\x10\xfd\x01\x12(\n" +
+	"#TRPC_STREAM_SERVER_READ_TIMEOUT_ERR\x10\xfe\x01\x12(\n" +
+	"#TRPC_STREAM_SERVER_IDLE_TIMEOUT_ERR\x10\xff\x01\x12#\n" +
+	"\x1eTRPC_STREAM_CLIENT_NETWORK_ERR\x10\xad\x02\x12,\n" +
+	"'TRPC_STREAM_CLIENT_MSG_EXCEED_LIMIT_ERR\x10\xb7\x02\x12\"\n" +
+	"\x1dTRPC_STREAM_CLIENT_ENCODE_ERR\x10\xc1\x02\x12\"\n" +
+	"\x1dTRPC_STREAM_CLIENT_DECODE_ERR\x10\xc2\x02\x12!\n" +
+	"\x1cTRPC_STREAM_CLIENT_WRITE_END\x10\xcb\x02\x12*\n" +
+	"%TRPC_STREAM_CLIENT_WRITE_OVERFLOW_ERR\x10\xcc\x02\x12'\n" +
+	"\"TRPC_STREAM_CLIENT_WRITE_CLOSE_ERR\x10\xcd\x02\x12)\n" +
+	"$TRPC_STREAM_CLIENT_WRITE_TIMEOUT_ERR\x10\xce\x02\x12 \n" +
+	"\x1bTRPC_STREAM_CLIENT_READ_END\x10\xdf\x02\x12&\n" +
+	"!TRPC_STREAM_CLIENT_READ_CLOSE_ERR\x10\xe0\x02\x12&\n" +
+	"!TRPC_STREAM_CLIENT_READ_EMPTY_ERR\x10\xe1\x02\x12(\n" +
+	"#TRPC_STREAM_CLIENT_READ_TIMEOUT_ERR\x10\xe2\x02\x12(\n" +
+	"#TRPC_STREAM_CLIENT_IDLE_TIMEOUT_ERR\x10\xe3\x02\x12\x1c\n" +
+	"\x17TRPC_INVOKE_UNKNOWN_ERR\x10\xe7\a\x12\x1c\n" +
+	"\x17TRPC_STREAM_UNKNOWN_ERR\x10\xe8\aB`\n" +
+	"&com.tencent.trpc.proto.standard.commonB\fTRPCProtocolZ(trpc.group/trpc/trpc-protocol/pb/go/trpcb\x06proto3"
 
 var (
 	file_trpc_trpc_proto_rawDescOnce sync.Once
-	file_trpc_trpc_proto_rawDescData = file_trpc_trpc_proto_rawDesc
+	file_trpc_trpc_proto_rawDescData []byte
 )
 
 func file_trpc_trpc_proto_rawDescGZIP() []byte {
 	file_trpc_trpc_proto_rawDescOnce.Do(func() {
-		file_trpc_trpc_proto_rawDescData = protoimpl.X.CompressGZIP(file_trpc_trpc_proto_rawDescData)
+		file_trpc_trpc_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_trpc_trpc_proto_rawDesc), len(file_trpc_trpc_proto_rawDesc)))
 	})
 	return file_trpc_trpc_proto_rawDescData
 }
 
 var file_trpc_trpc_proto_enumTypes = make([]protoimpl.EnumInfo, 10)
 var file_trpc_trpc_proto_msgTypes = make([]protoimpl.MessageInfo, 11)
-var file_trpc_trpc_proto_goTypes = []interface{}{
+var file_trpc_trpc_proto_goTypes = []any{
 	(TrpcMagic)(0),                     // 0: trpc.TrpcMagic
 	(TrpcDataFrameType)(0),             // 1: trpc.TrpcDataFrameType
 	(TrpcStreamFrameType)(0),           // 2: trpc.TrpcStreamFrameType
@@ -1887,97 +1725,11 @@ func file_trpc_trpc_proto_init() {
 	if File_trpc_trpc_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_trpc_trpc_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TrpcStreamInitMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_trpc_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TrpcStreamInitRequestMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_trpc_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TrpcStreamInitResponseMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_trpc_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TrpcStreamFeedBackMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_trpc_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TrpcStreamCloseMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_trpc_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RequestProtocol); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_trpc_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ResponseProtocol); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_trpc_trpc_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_trpc_trpc_proto_rawDesc), len(file_trpc_trpc_proto_rawDesc)),
 			NumEnums:      10,
 			NumMessages:   11,
 			NumExtensions: 0,
@@ -1989,7 +1741,6 @@ func file_trpc_trpc_proto_init() {
 		MessageInfos:      file_trpc_trpc_proto_msgTypes,
 	}.Build()
 	File_trpc_trpc_proto = out.File
-	file_trpc_trpc_proto_rawDesc = nil
 	file_trpc_trpc_proto_goTypes = nil
 	file_trpc_trpc_proto_depIdxs = nil
 }
