@@ -6,7 +6,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.36.6
 // 	protoc        v5.27.1
 // source: trpc/v2/api/http.proto
 
@@ -17,6 +17,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -27,10 +28,7 @@ const (
 )
 
 type HttpRule struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Selects a method to which this rule applies.
 	//
 	// Refer to [selector][google.api.DocumentationRule.selector] for syntax details.
@@ -39,7 +37,8 @@ type HttpRule struct {
 	// used with any of the {get|put|post|delete|patch} methods. A custom method
 	// can be defined using the 'custom' field.
 	//
-	// Types that are assignable to Pattern:
+	// Types that are valid to be assigned to Pattern:
+	//
 	//	*HttpRule_Get
 	//	*HttpRule_Put
 	//	*HttpRule_Post
@@ -65,15 +64,15 @@ type HttpRule struct {
 	// not contain an `additional_bindings` field themselves (that is,
 	// the nesting may only be one level deep).
 	AdditionalBindings []*HttpRule `protobuf:"bytes,11,rep,name=additional_bindings,json=additionalBindings,proto3" json:"additional_bindings,omitempty"`
+	unknownFields      protoimpl.UnknownFields
+	sizeCache          protoimpl.SizeCache
 }
 
 func (x *HttpRule) Reset() {
 	*x = HttpRule{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_v2_api_http_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_v2_api_http_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *HttpRule) String() string {
@@ -84,7 +83,7 @@ func (*HttpRule) ProtoMessage() {}
 
 func (x *HttpRule) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_v2_api_http_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -106,51 +105,63 @@ func (x *HttpRule) GetSelector() string {
 	return ""
 }
 
-func (m *HttpRule) GetPattern() isHttpRule_Pattern {
-	if m != nil {
-		return m.Pattern
+func (x *HttpRule) GetPattern() isHttpRule_Pattern {
+	if x != nil {
+		return x.Pattern
 	}
 	return nil
 }
 
 func (x *HttpRule) GetGet() string {
-	if x, ok := x.GetPattern().(*HttpRule_Get); ok {
-		return x.Get
+	if x != nil {
+		if x, ok := x.Pattern.(*HttpRule_Get); ok {
+			return x.Get
+		}
 	}
 	return ""
 }
 
 func (x *HttpRule) GetPut() string {
-	if x, ok := x.GetPattern().(*HttpRule_Put); ok {
-		return x.Put
+	if x != nil {
+		if x, ok := x.Pattern.(*HttpRule_Put); ok {
+			return x.Put
+		}
 	}
 	return ""
 }
 
 func (x *HttpRule) GetPost() string {
-	if x, ok := x.GetPattern().(*HttpRule_Post); ok {
-		return x.Post
+	if x != nil {
+		if x, ok := x.Pattern.(*HttpRule_Post); ok {
+			return x.Post
+		}
 	}
 	return ""
 }
 
 func (x *HttpRule) GetDelete() string {
-	if x, ok := x.GetPattern().(*HttpRule_Delete); ok {
-		return x.Delete
+	if x != nil {
+		if x, ok := x.Pattern.(*HttpRule_Delete); ok {
+			return x.Delete
+		}
 	}
 	return ""
 }
 
 func (x *HttpRule) GetPatch() string {
-	if x, ok := x.GetPattern().(*HttpRule_Patch); ok {
-		return x.Patch
+	if x != nil {
+		if x, ok := x.Pattern.(*HttpRule_Patch); ok {
+			return x.Patch
+		}
 	}
 	return ""
 }
 
 func (x *HttpRule) GetCustom() *CustomHttpPattern {
-	if x, ok := x.GetPattern().(*HttpRule_Custom); ok {
-		return x.Custom
+	if x != nil {
+		if x, ok := x.Pattern.(*HttpRule_Custom); ok {
+			return x.Custom
+		}
 	}
 	return nil
 }
@@ -228,23 +239,20 @@ func (*HttpRule_Custom) isHttpRule_Pattern() {}
 
 // A custom pattern is used for defining custom HTTP verb.
 type CustomHttpPattern struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// The name of this custom HTTP verb.
 	Kind string `protobuf:"bytes,1,opt,name=kind,proto3" json:"kind,omitempty"`
 	// The path matched by this custom verb.
-	Path string `protobuf:"bytes,2,opt,name=path,proto3" json:"path,omitempty"`
+	Path          string `protobuf:"bytes,2,opt,name=path,proto3" json:"path,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CustomHttpPattern) Reset() {
 	*x = CustomHttpPattern{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_v2_api_http_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_v2_api_http_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CustomHttpPattern) String() string {
@@ -255,7 +263,7 @@ func (*CustomHttpPattern) ProtoMessage() {}
 
 func (x *CustomHttpPattern) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_v2_api_http_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -286,57 +294,40 @@ func (x *CustomHttpPattern) GetPath() string {
 
 var File_trpc_v2_api_http_proto protoreflect.FileDescriptor
 
-var file_trpc_v2_api_http_proto_rawDesc = []byte{
-	0x0a, 0x16, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x76, 0x32, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x68, 0x74,
-	0x74, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0b, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76,
-	0x32, 0x2e, 0x61, 0x70, 0x69, 0x22, 0xdc, 0x02, 0x0a, 0x08, 0x48, 0x74, 0x74, 0x70, 0x52, 0x75,
-	0x6c, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x73, 0x65, 0x6c, 0x65, 0x63, 0x74, 0x6f, 0x72, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x65, 0x6c, 0x65, 0x63, 0x74, 0x6f, 0x72, 0x12, 0x12,
-	0x0a, 0x03, 0x67, 0x65, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x03, 0x67,
-	0x65, 0x74, 0x12, 0x12, 0x0a, 0x03, 0x70, 0x75, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x48,
-	0x00, 0x52, 0x03, 0x70, 0x75, 0x74, 0x12, 0x14, 0x0a, 0x04, 0x70, 0x6f, 0x73, 0x74, 0x18, 0x04,
-	0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x04, 0x70, 0x6f, 0x73, 0x74, 0x12, 0x18, 0x0a, 0x06,
-	0x64, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x06,
-	0x64, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x12, 0x16, 0x0a, 0x05, 0x70, 0x61, 0x74, 0x63, 0x68, 0x18,
-	0x06, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x05, 0x70, 0x61, 0x74, 0x63, 0x68, 0x12, 0x38,
-	0x0a, 0x06, 0x63, 0x75, 0x73, 0x74, 0x6f, 0x6d, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1e,
-	0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x32, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x43, 0x75, 0x73,
-	0x74, 0x6f, 0x6d, 0x48, 0x74, 0x74, 0x70, 0x50, 0x61, 0x74, 0x74, 0x65, 0x72, 0x6e, 0x48, 0x00,
-	0x52, 0x06, 0x63, 0x75, 0x73, 0x74, 0x6f, 0x6d, 0x12, 0x12, 0x0a, 0x04, 0x62, 0x6f, 0x64, 0x79,
-	0x18, 0x07, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x62, 0x6f, 0x64, 0x79, 0x12, 0x23, 0x0a, 0x0d,
-	0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x5f, 0x62, 0x6f, 0x64, 0x79, 0x18, 0x0c, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x0c, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x6f, 0x64,
-	0x79, 0x12, 0x46, 0x0a, 0x13, 0x61, 0x64, 0x64, 0x69, 0x74, 0x69, 0x6f, 0x6e, 0x61, 0x6c, 0x5f,
-	0x62, 0x69, 0x6e, 0x64, 0x69, 0x6e, 0x67, 0x73, 0x18, 0x0b, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x15,
-	0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x32, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x48, 0x74, 0x74,
-	0x70, 0x52, 0x75, 0x6c, 0x65, 0x52, 0x12, 0x61, 0x64, 0x64, 0x69, 0x74, 0x69, 0x6f, 0x6e, 0x61,
-	0x6c, 0x42, 0x69, 0x6e, 0x64, 0x69, 0x6e, 0x67, 0x73, 0x42, 0x09, 0x0a, 0x07, 0x70, 0x61, 0x74,
-	0x74, 0x65, 0x72, 0x6e, 0x22, 0x3b, 0x0a, 0x11, 0x43, 0x75, 0x73, 0x74, 0x6f, 0x6d, 0x48, 0x74,
-	0x74, 0x70, 0x50, 0x61, 0x74, 0x74, 0x65, 0x72, 0x6e, 0x12, 0x12, 0x0a, 0x04, 0x6b, 0x69, 0x6e,
-	0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6b, 0x69, 0x6e, 0x64, 0x12, 0x12, 0x0a,
-	0x04, 0x70, 0x61, 0x74, 0x68, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x70, 0x61, 0x74,
-	0x68, 0x42, 0x50, 0x0a, 0x1d, 0x63, 0x6f, 0x6d, 0x2e, 0x74, 0x65, 0x6e, 0x63, 0x65, 0x6e, 0x74,
-	0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x65,
-	0x78, 0x74, 0x5a, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x67, 0x72, 0x6f, 0x75, 0x70, 0x2f, 0x74,
-	0x72, 0x70, 0x63, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f,
-	0x6c, 0x2f, 0x70, 0x62, 0x2f, 0x67, 0x6f, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x76, 0x32, 0x2f,
-	0x61, 0x70, 0x69, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_trpc_v2_api_http_proto_rawDesc = "" +
+	"\n" +
+	"\x16trpc/v2/api/http.proto\x12\vtrpc.v2.api\"\xdc\x02\n" +
+	"\bHttpRule\x12\x1a\n" +
+	"\bselector\x18\x01 \x01(\tR\bselector\x12\x12\n" +
+	"\x03get\x18\x02 \x01(\tH\x00R\x03get\x12\x12\n" +
+	"\x03put\x18\x03 \x01(\tH\x00R\x03put\x12\x14\n" +
+	"\x04post\x18\x04 \x01(\tH\x00R\x04post\x12\x18\n" +
+	"\x06delete\x18\x05 \x01(\tH\x00R\x06delete\x12\x16\n" +
+	"\x05patch\x18\x06 \x01(\tH\x00R\x05patch\x128\n" +
+	"\x06custom\x18\b \x01(\v2\x1e.trpc.v2.api.CustomHttpPatternH\x00R\x06custom\x12\x12\n" +
+	"\x04body\x18\a \x01(\tR\x04body\x12#\n" +
+	"\rresponse_body\x18\f \x01(\tR\fresponseBody\x12F\n" +
+	"\x13additional_bindings\x18\v \x03(\v2\x15.trpc.v2.api.HttpRuleR\x12additionalBindingsB\t\n" +
+	"\apattern\";\n" +
+	"\x11CustomHttpPattern\x12\x12\n" +
+	"\x04kind\x18\x01 \x01(\tR\x04kind\x12\x12\n" +
+	"\x04path\x18\x02 \x01(\tR\x04pathBP\n" +
+	"\x1dcom.tencent.trpc.protobuf.extZ/trpc.group/trpc/trpc-protocol/pb/go/trpc/v2/apib\x06proto3"
 
 var (
 	file_trpc_v2_api_http_proto_rawDescOnce sync.Once
-	file_trpc_v2_api_http_proto_rawDescData = file_trpc_v2_api_http_proto_rawDesc
+	file_trpc_v2_api_http_proto_rawDescData []byte
 )
 
 func file_trpc_v2_api_http_proto_rawDescGZIP() []byte {
 	file_trpc_v2_api_http_proto_rawDescOnce.Do(func() {
-		file_trpc_v2_api_http_proto_rawDescData = protoimpl.X.CompressGZIP(file_trpc_v2_api_http_proto_rawDescData)
+		file_trpc_v2_api_http_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_trpc_v2_api_http_proto_rawDesc), len(file_trpc_v2_api_http_proto_rawDesc)))
 	})
 	return file_trpc_v2_api_http_proto_rawDescData
 }
 
 var file_trpc_v2_api_http_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_trpc_v2_api_http_proto_goTypes = []interface{}{
+var file_trpc_v2_api_http_proto_goTypes = []any{
 	(*HttpRule)(nil),          // 0: trpc.v2.api.HttpRule
 	(*CustomHttpPattern)(nil), // 1: trpc.v2.api.CustomHttpPattern
 }
@@ -355,33 +346,7 @@ func file_trpc_v2_api_http_proto_init() {
 	if File_trpc_v2_api_http_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_trpc_v2_api_http_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*HttpRule); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_v2_api_http_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CustomHttpPattern); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_trpc_v2_api_http_proto_msgTypes[0].OneofWrappers = []interface{}{
+	file_trpc_v2_api_http_proto_msgTypes[0].OneofWrappers = []any{
 		(*HttpRule_Get)(nil),
 		(*HttpRule_Put)(nil),
 		(*HttpRule_Post)(nil),
@@ -393,7 +358,7 @@ func file_trpc_v2_api_http_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_trpc_v2_api_http_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_trpc_v2_api_http_proto_rawDesc), len(file_trpc_v2_api_http_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -404,7 +369,6 @@ func file_trpc_v2_api_http_proto_init() {
 		MessageInfos:      file_trpc_v2_api_http_proto_msgTypes,
 	}.Build()
 	File_trpc_v2_api_http_proto = out.File
-	file_trpc_v2_api_http_proto_rawDesc = nil
 	file_trpc_v2_api_http_proto_goTypes = nil
 	file_trpc_v2_api_http_proto_depIdxs = nil
 }
