@@ -6,7 +6,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.36.6
 // 	protoc        v5.27.1
 // source: trpc/v2/proto/trpc_options.proto
 
@@ -17,6 +17,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	descriptorpb "google.golang.org/protobuf/types/descriptorpb"
 	reflect "reflect"
+	unsafe "unsafe"
 )
 
 const (
@@ -59,28 +60,14 @@ var (
 
 var File_trpc_v2_proto_trpc_options_proto protoreflect.FileDescriptor
 
-var file_trpc_v2_proto_trpc_options_proto_rawDesc = []byte{
-	0x0a, 0x20, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x76, 0x32, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f,
-	0x74, 0x72, 0x70, 0x63, 0x5f, 0x6f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x12, 0x07, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x32, 0x1a, 0x20, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x64, 0x65, 0x73,
-	0x63, 0x72, 0x69, 0x70, 0x74, 0x6f, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x3a, 0x35, 0x0a,
-	0x05, 0x61, 0x6c, 0x69, 0x61, 0x73, 0x12, 0x1e, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x4f,
-	0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x18, 0x83, 0x09, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x61,
-	0x6c, 0x69, 0x61, 0x73, 0x3a, 0x35, 0x0a, 0x06, 0x67, 0x6f, 0x5f, 0x74, 0x61, 0x67, 0x12, 0x1d,
-	0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66,
-	0x2e, 0x46, 0x69, 0x65, 0x6c, 0x64, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x18, 0x84, 0x09,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x67, 0x6f, 0x54, 0x61, 0x67, 0x42, 0x52, 0x0a, 0x1d, 0x63,
-	0x6f, 0x6d, 0x2e, 0x74, 0x65, 0x6e, 0x63, 0x65, 0x6e, 0x74, 0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x65, 0x78, 0x74, 0x5a, 0x31, 0x74, 0x72,
-	0x70, 0x63, 0x2e, 0x67, 0x72, 0x6f, 0x75, 0x70, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x74, 0x72,
-	0x70, 0x63, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x2f, 0x70, 0x62, 0x2f, 0x67,
-	0x6f, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x76, 0x32, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_trpc_v2_proto_trpc_options_proto_rawDesc = "" +
+	"\n" +
+	" trpc/v2/proto/trpc_options.proto\x12\atrpc.v2\x1a google/protobuf/descriptor.proto:5\n" +
+	"\x05alias\x12\x1e.google.protobuf.MethodOptions\x18\x83\t \x01(\tR\x05alias:5\n" +
+	"\x06go_tag\x12\x1d.google.protobuf.FieldOptions\x18\x84\t \x01(\tR\x05goTagBR\n" +
+	"\x1dcom.tencent.trpc.protobuf.extZ1trpc.group/trpc/trpc-protocol/pb/go/trpc/v2/protob\x06proto3"
 
-var file_trpc_v2_proto_trpc_options_proto_goTypes = []interface{}{
+var file_trpc_v2_proto_trpc_options_proto_goTypes = []any{
 	(*descriptorpb.MethodOptions)(nil), // 0: google.protobuf.MethodOptions
 	(*descriptorpb.FieldOptions)(nil),  // 1: google.protobuf.FieldOptions
 }
@@ -103,7 +90,7 @@ func file_trpc_v2_proto_trpc_options_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_trpc_v2_proto_trpc_options_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_trpc_v2_proto_trpc_options_proto_rawDesc), len(file_trpc_v2_proto_trpc_options_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   0,
 			NumExtensions: 2,
@@ -114,7 +101,6 @@ func file_trpc_v2_proto_trpc_options_proto_init() {
 		ExtensionInfos:    file_trpc_v2_proto_trpc_options_proto_extTypes,
 	}.Build()
 	File_trpc_v2_proto_trpc_options_proto = out.File
-	file_trpc_v2_proto_trpc_options_proto_rawDesc = nil
 	file_trpc_v2_proto_trpc_options_proto_goTypes = nil
 	file_trpc_v2_proto_trpc_options_proto_depIdxs = nil
 }
