@@ -6,7 +6,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.36.6
 // 	protoc        v5.27.1
 // source: trpc/v2/swagger/swagger.proto
 
@@ -18,6 +18,7 @@ import (
 	descriptorpb "google.golang.org/protobuf/types/descriptorpb"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -29,23 +30,20 @@ const (
 
 // to gen swagger json
 type SwaggerRule struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Title         string                 `protobuf:"bytes,50103,opt,name=title,proto3" json:"title,omitempty"`
+	Method        string                 `protobuf:"bytes,50104,opt,name=method,proto3" json:"method,omitempty"`
+	Description   string                 `protobuf:"bytes,50105,opt,name=description,proto3" json:"description,omitempty"`
+	Params        []*SwaggerParam        `protobuf:"bytes,50106,rep,name=params,proto3" json:"params,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Title       string          `protobuf:"bytes,50103,opt,name=title,proto3" json:"title,omitempty"`
-	Method      string          `protobuf:"bytes,50104,opt,name=method,proto3" json:"method,omitempty"`
-	Description string          `protobuf:"bytes,50105,opt,name=description,proto3" json:"description,omitempty"`
-	Params      []*SwaggerParam `protobuf:"bytes,50106,rep,name=params,proto3" json:"params,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SwaggerRule) Reset() {
 	*x = SwaggerRule{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_v2_swagger_swagger_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_v2_swagger_swagger_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SwaggerRule) String() string {
@@ -56,7 +54,7 @@ func (*SwaggerRule) ProtoMessage() {}
 
 func (x *SwaggerRule) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_v2_swagger_swagger_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -100,22 +98,19 @@ func (x *SwaggerRule) GetParams() []*SwaggerParam {
 }
 
 type SwaggerParam struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Name          string                 `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
+	Required      bool                   `protobuf:"varint,2,opt,name=required,proto3" json:"required,omitempty"`
+	Default       string                 `protobuf:"bytes,3,opt,name=default,proto3" json:"default,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Name     string `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
-	Required bool   `protobuf:"varint,2,opt,name=required,proto3" json:"required,omitempty"`
-	Default  string `protobuf:"bytes,3,opt,name=default,proto3" json:"default,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SwaggerParam) Reset() {
 	*x = SwaggerParam{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_trpc_v2_swagger_swagger_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_trpc_v2_swagger_swagger_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SwaggerParam) String() string {
@@ -126,7 +121,7 @@ func (*SwaggerParam) ProtoMessage() {}
 
 func (x *SwaggerParam) ProtoReflect() protoreflect.Message {
 	mi := &file_trpc_v2_swagger_swagger_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -181,54 +176,35 @@ var (
 
 var File_trpc_v2_swagger_swagger_proto protoreflect.FileDescriptor
 
-var file_trpc_v2_swagger_swagger_proto_rawDesc = []byte{
-	0x0a, 0x1d, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x76, 0x32, 0x2f, 0x73, 0x77, 0x61, 0x67, 0x67, 0x65,
-	0x72, 0x2f, 0x73, 0x77, 0x61, 0x67, 0x67, 0x65, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
-	0x07, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x32, 0x1a, 0x20, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
-	0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x64, 0x65, 0x73, 0x63, 0x72, 0x69,
-	0x70, 0x74, 0x6f, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x94, 0x01, 0x0a, 0x0b, 0x53,
-	0x77, 0x61, 0x67, 0x67, 0x65, 0x72, 0x52, 0x75, 0x6c, 0x65, 0x12, 0x16, 0x0a, 0x05, 0x74, 0x69,
-	0x74, 0x6c, 0x65, 0x18, 0xb7, 0x87, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x74, 0x69, 0x74,
-	0x6c, 0x65, 0x12, 0x18, 0x0a, 0x06, 0x6d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x18, 0xb8, 0x87, 0x03,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x6d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x12, 0x22, 0x0a, 0x0b,
-	0x64, 0x65, 0x73, 0x63, 0x72, 0x69, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0xb9, 0x87, 0x03, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x0b, 0x64, 0x65, 0x73, 0x63, 0x72, 0x69, 0x70, 0x74, 0x69, 0x6f, 0x6e,
-	0x12, 0x2f, 0x0a, 0x06, 0x70, 0x61, 0x72, 0x61, 0x6d, 0x73, 0x18, 0xba, 0x87, 0x03, 0x20, 0x03,
-	0x28, 0x0b, 0x32, 0x15, 0x2e, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x32, 0x2e, 0x53, 0x77, 0x61,
-	0x67, 0x67, 0x65, 0x72, 0x50, 0x61, 0x72, 0x61, 0x6d, 0x52, 0x06, 0x70, 0x61, 0x72, 0x61, 0x6d,
-	0x73, 0x22, 0x58, 0x0a, 0x0c, 0x53, 0x77, 0x61, 0x67, 0x67, 0x65, 0x72, 0x50, 0x61, 0x72, 0x61,
-	0x6d, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x04, 0x6e, 0x61, 0x6d, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x72, 0x65, 0x71, 0x75, 0x69, 0x72, 0x65,
-	0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x08, 0x52, 0x08, 0x72, 0x65, 0x71, 0x75, 0x69, 0x72, 0x65,
-	0x64, 0x12, 0x18, 0x0a, 0x07, 0x64, 0x65, 0x66, 0x61, 0x75, 0x6c, 0x74, 0x18, 0x03, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x07, 0x64, 0x65, 0x66, 0x61, 0x75, 0x6c, 0x74, 0x3a, 0x4f, 0x0a, 0x07, 0x73,
-	0x77, 0x61, 0x67, 0x67, 0x65, 0x72, 0x12, 0x1e, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x4f,
-	0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x18, 0x84, 0x09, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x14, 0x2e,
-	0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x32, 0x2e, 0x53, 0x77, 0x61, 0x67, 0x67, 0x65, 0x72, 0x52,
-	0x75, 0x6c, 0x65, 0x52, 0x07, 0x73, 0x77, 0x61, 0x67, 0x67, 0x65, 0x72, 0x42, 0x54, 0x0a, 0x1d,
-	0x63, 0x6f, 0x6d, 0x2e, 0x74, 0x65, 0x6e, 0x63, 0x65, 0x6e, 0x74, 0x2e, 0x74, 0x72, 0x70, 0x63,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x65, 0x78, 0x74, 0x5a, 0x33, 0x74,
-	0x72, 0x70, 0x63, 0x2e, 0x67, 0x72, 0x6f, 0x75, 0x70, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x74,
-	0x72, 0x70, 0x63, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x2f, 0x70, 0x62, 0x2f,
-	0x67, 0x6f, 0x2f, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x76, 0x32, 0x2f, 0x73, 0x77, 0x61, 0x67, 0x67,
-	0x65, 0x72, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_trpc_v2_swagger_swagger_proto_rawDesc = "" +
+	"\n" +
+	"\x1dtrpc/v2/swagger/swagger.proto\x12\atrpc.v2\x1a google/protobuf/descriptor.proto\"\x94\x01\n" +
+	"\vSwaggerRule\x12\x16\n" +
+	"\x05title\x18\xb7\x87\x03 \x01(\tR\x05title\x12\x18\n" +
+	"\x06method\x18\xb8\x87\x03 \x01(\tR\x06method\x12\"\n" +
+	"\vdescription\x18\xb9\x87\x03 \x01(\tR\vdescription\x12/\n" +
+	"\x06params\x18\xba\x87\x03 \x03(\v2\x15.trpc.v2.SwaggerParamR\x06params\"X\n" +
+	"\fSwaggerParam\x12\x12\n" +
+	"\x04name\x18\x01 \x01(\tR\x04name\x12\x1a\n" +
+	"\brequired\x18\x02 \x01(\bR\brequired\x12\x18\n" +
+	"\adefault\x18\x03 \x01(\tR\adefault:O\n" +
+	"\aswagger\x12\x1e.google.protobuf.MethodOptions\x18\x84\t \x01(\v2\x14.trpc.v2.SwaggerRuleR\aswaggerBT\n" +
+	"\x1dcom.tencent.trpc.protobuf.extZ3trpc.group/trpc/trpc-protocol/pb/go/trpc/v2/swaggerb\x06proto3"
 
 var (
 	file_trpc_v2_swagger_swagger_proto_rawDescOnce sync.Once
-	file_trpc_v2_swagger_swagger_proto_rawDescData = file_trpc_v2_swagger_swagger_proto_rawDesc
+	file_trpc_v2_swagger_swagger_proto_rawDescData []byte
 )
 
 func file_trpc_v2_swagger_swagger_proto_rawDescGZIP() []byte {
 	file_trpc_v2_swagger_swagger_proto_rawDescOnce.Do(func() {
-		file_trpc_v2_swagger_swagger_proto_rawDescData = protoimpl.X.CompressGZIP(file_trpc_v2_swagger_swagger_proto_rawDescData)
+		file_trpc_v2_swagger_swagger_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_trpc_v2_swagger_swagger_proto_rawDesc), len(file_trpc_v2_swagger_swagger_proto_rawDesc)))
 	})
 	return file_trpc_v2_swagger_swagger_proto_rawDescData
 }
 
 var file_trpc_v2_swagger_swagger_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_trpc_v2_swagger_swagger_proto_goTypes = []interface{}{
+var file_trpc_v2_swagger_swagger_proto_goTypes = []any{
 	(*SwaggerRule)(nil),                // 0: trpc.v2.SwaggerRule
 	(*SwaggerParam)(nil),               // 1: trpc.v2.SwaggerParam
 	(*descriptorpb.MethodOptions)(nil), // 2: google.protobuf.MethodOptions
@@ -249,37 +225,11 @@ func file_trpc_v2_swagger_swagger_proto_init() {
 	if File_trpc_v2_swagger_swagger_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_trpc_v2_swagger_swagger_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SwaggerRule); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_trpc_v2_swagger_swagger_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SwaggerParam); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_trpc_v2_swagger_swagger_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_trpc_v2_swagger_swagger_proto_rawDesc), len(file_trpc_v2_swagger_swagger_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 1,
@@ -291,7 +241,6 @@ func file_trpc_v2_swagger_swagger_proto_init() {
 		ExtensionInfos:    file_trpc_v2_swagger_swagger_proto_extTypes,
 	}.Build()
 	File_trpc_v2_swagger_swagger_proto = out.File
-	file_trpc_v2_swagger_swagger_proto_rawDesc = nil
 	file_trpc_v2_swagger_swagger_proto_goTypes = nil
 	file_trpc_v2_swagger_swagger_proto_depIdxs = nil
 }
